Code resolve
