--- docs/docs/auto-docs/components/UserPortal/ChatRoom/mocks/variables/CHATS_LIST_MOCK.md
@@ -1,9 +1,0 @@
-[Admin Docs](/)
-
-***
-
-# Variable: CHATS\_LIST\_MOCK
-
-> `const` **CHATS\_LIST\_MOCK**: (\{ `request`: \{ `query`: `DocumentNode`; `variables`: \{ `id`: `string`; \}; \}; `result`: \{ `data`: \{ `chatsByUserId`: (\{ `_id`: `string`; `admins`: `object`[]; `createdAt`: `string`; `creator`: \{ `__typename`: `string`; `_id`: `string`; `createdAt`: `string`; `email`: `string`; `firstName`: `string`; `image`: `any`; `lastName`: `string`; \}; `isGroup`: `boolean`; `messages`: `object`[]; `name`: `string`; `organization`: \{ `_id`: `string`; `name`: `string`; \}; `unseenMessagesByUsers`: \{ `1`: `number`; `2`: `number`; `3`: `number`; `4`: `number`; `5`: `number`; \}; `users`: `object`[]; \} \| \{ `_id`: `string`; `admins`: `undefined`; `createdAt`: `string`; `creator`: \{ `__typename`: `string`; `_id`: `string`; `createdAt`: `string`; `email`: `string`; `firstName`: `string`; `image`: `any`; `lastName`: `string`; \}; `isGroup`: `boolean`; `messages`: `object`[]; `name`: `string`; `organization`: \{ `_id`: `string`; `name`: `string`; \}; `unseenMessagesByUsers`: \{ `1`: `number`; `2`: `number`; `3`: `number`; `4`: `number`; `5`: `number`; \}; `users`: `object`[]; \})[]; \}; \}; \} \| \{ `request`: \{ `query`: `DocumentNode`; `variables`: \{ `id`: `string`; \}; \}; `result`: \{ `data`: \{ `chatsByUserId`: `object`[]; \}; \}; \} \| \{ `request`: \{ `query`: `DocumentNode`; `variables`: \{ `id`: `string`; \}; \}; `result`: \{ `data`: \{ `chatsByUserId`: `object`[]; \}; \}; \} \| \{ `request`: \{ `query`: `DocumentNode`; `variables`: \{ `id`: `string`; \}; \}; `result`: \{ `data`: \{ `chatsByUserId`: (\{ `_id`: `string`; `admins`: `object`[]; `createdAt`: `string`; `creator`: \{ `__typename`: `string`; `_id`: `string`; `createdAt`: `string`; `email`: `string`; `firstName`: `string`; `image`: `any`; `lastName`: `string`; \}; `isGroup`: `boolean`; `messages`: `object`[]; `name`: `string`; `organization`: \{ `_id`: `string`; `name`: `string`; \}; `unseenMessagesByUsers`: \{ `1`: `number`; `2`: `number`; `3`: `number`; `4`: `number`; `5`: `number`; \}; `users`: `object`[]; \} \| \{ `_id`: `string`; `admins`: `object`[]; `createdAt`: `string`; `creator`: \{ `__typename`: `string`; `_id`: `string`; `createdAt`: `string`; `email`: `string`; `firstName`: `string`; `image`: `any`; `lastName`: `string`; \}; `isGroup`: `boolean`; `messages`: `object`[]; `name`: `string`; `organization`: \{ `_id`: `string`; `name`: `string`; \}; `unseenMessagesByUsers`: `string`; `users`: `object`[]; \})[]; \}; \}; \})[]
-
-Defined in: [src/components/UserPortal/ChatRoom/mocks.ts:2052](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/components/UserPortal/ChatRoom/mocks.ts#L2052)
