--- docs/docs/auto-docs/components/UserPortal/ChatRoom/mocks/variables/MARK_CHAT_MESSAGES_AS_READ_MOCK.md
@@ -1,9 +1,0 @@
-[Admin Docs](/)
-
-***
-
-# Variable: MARK\_CHAT\_MESSAGES\_AS\_READ\_MOCK
-
-> `const` **MARK\_CHAT\_MESSAGES\_AS\_READ\_MOCK**: `any`[]
-
-Defined in: [src/components/UserPortal/ChatRoom/mocks.ts:5226](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/components/UserPortal/ChatRoom/mocks.ts#L5226)
