--- docs/docs/auto-docs/components/UserPortal/ChatRoom/mocks/variables/MESSAGE_SENT_TO_CHAT_MOCK.md
@@ -1,91 +1,0 @@
-[Admin Docs](/)
-
-***
-
-# Variable: MESSAGE\_SENT\_TO\_CHAT\_MOCK
-
-> `const` **MESSAGE\_SENT\_TO\_CHAT\_MOCK**: `object`[]
-
-Defined in: [src/components/UserPortal/ChatRoom/mocks.ts:13](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/components/UserPortal/ChatRoom/mocks.ts#L13)
-
-## Type declaration
-
-### request
-
-> **request**: `object`
-
-#### request.query
-
-> **query**: `DocumentNode` = `MESSAGE_SENT_TO_CHAT`
-
-#### request.variables
-
-> **variables**: `object`
-
-#### request.variables.userId
-
-> **userId**: `string` = `'2'`
-
-### result
-
-> **result**: `object`
-
-#### result.data
-
-> **data**: `object`
-
-#### result.data.messageSentToChat
-
-> **messageSentToChat**: `object`
-
-#### result.data.messageSentToChat.\_id
-
-> **\_id**: `string` = `'668ec1f1df364e03ac47a151'`
-
-#### result.data.messageSentToChat.chatMessageBelongsTo
-
-> **chatMessageBelongsTo**: `object`
-
-#### result.data.messageSentToChat.chatMessageBelongsTo.\_id
-
-> **\_id**: `string` = `'1'`
-
-#### result.data.messageSentToChat.createdAt
-
-> **createdAt**: `string` = `'2024-07-10T17:16:33.248Z'`
-
-#### result.data.messageSentToChat.media
-
-> **media**: `any` = `null`
-
-#### result.data.messageSentToChat.messageContent
-
-> **messageContent**: `string` = `'Test '`
-
-#### result.data.messageSentToChat.replyTo
-
-> **replyTo**: `any` = `null`
-
-#### result.data.messageSentToChat.sender
-
-> **sender**: `object`
-
-#### result.data.messageSentToChat.sender.\_id
-
-> **\_id**: `string` = `'64378abd85008f171cf2990d'`
-
-#### result.data.messageSentToChat.sender.firstName
-
-> **firstName**: `string` = `'Wilt'`
-
-#### result.data.messageSentToChat.sender.image
-
-> **image**: `string` = `''`
-
-#### result.data.messageSentToChat.sender.lastName
-
-> **lastName**: `string` = `'Shepherd'`
-
-#### result.data.messageSentToChat.updatedAt
-
-> **updatedAt**: `string` = `'2024-07-10'`
