--- docs/docs/auto-docs/components/UserPortal/ChatRoom/mocks/variables/SEND_MESSAGE_TO_CHAT_MOCK.md
@@ -1,9 +1,0 @@
-[Admin Docs](/)
-
-***
-
-# Variable: SEND\_MESSAGE\_TO\_CHAT\_MOCK
-
-> `const` **SEND\_MESSAGE\_TO\_CHAT\_MOCK**: (\{ `request`: \{ `query`: `DocumentNode`; `variables`: \{ `chatId`: `string`; `media`: `string`; `messageContent`: `string`; `replyTo`: `string`; \}; \}; `result`: \{ `data`: \{ `sendMessageToChat`: \{ `_id`: `string`; `createdAt`: `string`; `media`: `string`; `messageContent`: `string`; `replyTo`: `any`; `sender`: \{ `_id`: `string`; `firstName`: `string`; `image`: `string`; `lastName`: `string`; \}; `updatedAt`: `string`; \}; \}; \}; \} \| \{ `request`: \{ `query`: `DocumentNode`; `variables`: \{ `chatId`: `string`; `media`: `undefined`; `messageContent`: `string`; `replyTo`: `string`; \}; \}; `result`: \{ `data`: \{ `sendMessageToChat`: \{ `_id`: `string`; `createdAt`: `string`; `media`: `any`; `messageContent`: `string`; `replyTo`: `any`; `sender`: \{ `_id`: `string`; `firstName`: `string`; `image`: `string`; `lastName`: `string`; \}; `updatedAt`: `string`; \}; \}; \}; \})[]
-
-Defined in: [src/components/UserPortal/ChatRoom/mocks.ts:5021](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/components/UserPortal/ChatRoom/mocks.ts#L5021)
