--- docs/docs/auto-docs/components/UserPortal/ChatRoom/mocks/variables/UNREAD_CHAT_LIST_QUERY_MOCK.md
@@ -1,9 +1,0 @@
-[Admin Docs](/)
-
-***
-
-# Variable: UNREAD\_CHAT\_LIST\_QUERY\_MOCK
-
-> `const` **UNREAD\_CHAT\_LIST\_QUERY\_MOCK**: (\{ `request`: \{ `query`: `DocumentNode`; `variables`: \{ `id`: `undefined`; \}; \}; `result`: \{ `data`: \{ `getUnreadChatsByUserId`: \{ `_id`: `string`; `admins`: `any`[]; `createdAt`: `string`; `creator`: \{ `__typename`: `string`; `_id`: `string`; `createdAt`: `string`; `email`: `string`; `firstName`: `string`; `image`: `any`; `lastName`: `string`; \}; `isGroup`: `boolean`; `messages`: `object`[]; `name`: `string`; `organization`: `any`; `unseenMessagesByUsers`: `string`; `users`: `object`[]; \}; \}; \}; \} \| \{ `request`: \{ `query`: `DocumentNode`; `variables`: \{ `id`: `string`; \}; \}; `result`: \{ `data`: \{ `getUnreadChatsByUserId`: \{ `_id`: `string`; `admins`: `any`[]; `createdAt`: `string`; `creator`: \{ `__typename`: `string`; `_id`: `string`; `createdAt`: `string`; `email`: `string`; `firstName`: `string`; `image`: `any`; `lastName`: `string`; \}; `isGroup`: `boolean`; `messages`: `object`[]; `name`: `string`; `organization`: `any`; `unseenMessagesByUsers`: `string`; `users`: `object`[]; \}; \}; \}; \})[]
-
-Defined in: [src/components/UserPortal/ChatRoom/mocks.ts:132](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/components/UserPortal/ChatRoom/mocks.ts#L132)
