+++ eslint.config.js
@@ -1,41 +1,41 @@
-import js from "@eslint/js";
-import ts from "@typescript-eslint/eslint-plugin";
-import tsParser from "@typescript-eslint/parser";
-import react from "eslint-plugin-react";
-import prettier from "eslint-plugin-prettier";
-import vitest from "eslint-plugin-vitest";
-import tsdoc from "eslint-plugin-tsdoc";
-import graphql from "@graphql-eslint/eslint-plugin";
-import imports from "eslint-plugin-import";
+import js from '@eslint/js';
+import ts from '@typescript-eslint/eslint-plugin';
+import tsParser from '@typescript-eslint/parser';
+import react from 'eslint-plugin-react';
+import prettier from 'eslint-plugin-prettier';
+import vitest from 'eslint-plugin-vitest';
+import tsdoc from 'eslint-plugin-tsdoc';
+import graphql from '@graphql-eslint/eslint-plugin';
+import imports from 'eslint-plugin-import';
 
 export default [
   {
     ignores: [
-      "node_modules",
-      "dist",
-      "build",
-      "package.json",
-      "package-lock.json",
-      "tsconfig.json",
-      "fix-readme-links.js",
-      "fix-repo-url.js",
-      "src/components/CheckIn/tagTemplate.ts",
-      "docs/**",
-      "*.md",
-      "docker/**",
-      "config/docker/setup/nginx*.conf",
-      "**/*.css",
-      "**/*.scss",
-      "**/*.less",
-      "**/*.json",
-      "**/*.svg"
+      'node_modules',
+      'dist',
+      'build',
+      'package.json',
+      'package-lock.json',
+      'tsconfig.json',
+      'fix-readme-links.js',
+      'fix-repo-url.js',
+      'src/components/CheckIn/tagTemplate.ts',
+      'docs/**',
+      '*.md',
+      'docker/**',
+      'config/docker/setup/nginx*.conf',
+      '**/*.css',
+      '**/*.scss',
+      '**/*.less',
+      '**/*.json',
+      '**/*.svg',
     ],
   },
   {
-    files: ["**/*.ts", "**/*.tsx"],
+    files: ['**/*.ts', '**/*.tsx'],
     languageOptions: {
-      ecmaVersion: "latest",
-      sourceType: "module",
+      ecmaVersion: 'latest',
+      sourceType: 'module',
       parser: tsParser,
       parserOptions: {
         ecmaFeatures: {
@@ -44,85 +44,105 @@ export default [
       },
       globals: {
         // ✅ Node.js globals
-        window: "readonly",
-        localStorage: "readonly",
-        setTimeout: "readonly",
-        console: "readonly",
+        window: 'readonly',
+        localStorage: 'readonly',
+        setTimeout: 'readonly',
+        console: 'readonly',
+        HTMLImageElement: 'readonly',
+        HTMLInputElement: 'readonly',
+        HTMLInputElement: 'readonly',
+        File: 'readonly',
 
         // ✅ Vitest globals
-        describe: "readonly",
-        test: "readonly",
-        expect: "readonly",
-        beforeEach: "readonly",
-        afterEach: "readonly",
-        beforeAll: "readonly",
-        afterAll: "readonly",
+        describe: 'readonly',
+        test: 'readonly',
+        expect: 'readonly',
+        beforeEach: 'readonly',
+        afterEach: 'readonly',
+        beforeAll: 'readonly',
+        afterAll: 'readonly',
       },
     },
     plugins: {
       react,
-      "@typescript-eslint": ts,
+      '@typescript-eslint': ts,
       vitest,
       tsdoc,
-      "@graphql-eslint": graphql,
+      '@graphql-eslint': graphql,
       import: imports,
       prettier,
     },
     settings: {
       react: {
-        version: "detect",
+        version: 'detect',
       },
     },
     rules: {
       ...js.configs.recommended.rules,
       ...ts.configs.recommended.rules,
-      "react/destructuring-assignment": "error",
-      "react/no-multi-comp": ["error", { ignoreStateless: false }],
-      "react/jsx-filename-extension": ["error", { extensions: [".tsx"] }],
-      "import/no-duplicates": "error",
-      "tsdoc/syntax": "error",
-      "@typescript-eslint/ban-ts-comment": "error",
-      "@typescript-eslint/no-unused-vars": "error",
-      "@typescript-eslint/no-explicit-any": "error",
-      "@typescript-eslint/no-non-null-assertion": "error",
-      "@typescript-eslint/consistent-type-assertions": "error",
-      "@typescript-eslint/naming-convention": [
-        "error",
-        { selector: "interface", format: ["PascalCase"], prefix: ["I"] },
-        { selector: ["typeAlias", "typeLike", "enum"], format: ["PascalCase"] },
-        { selector: "typeParameter", format: ["PascalCase"], prefix: ["T"] },
-        { selector: "variable", format: ["camelCase", "UPPER_CASE", "PascalCase"], leadingUnderscore: "allow" },
-        { selector: "parameter", format: ["camelCase"], leadingUnderscore: "allow" },
-        { selector: "function", format: ["camelCase", "PascalCase"] },
-        { selector: "memberLike", modifiers: ["private"], format: ["camelCase"], leadingUnderscore: "require" },
-        { selector: "variable", modifiers: ["exported"], format: null },
+      'react/destructuring-assignment': 'error',
+      'react/no-multi-comp': ['error', { ignoreStateless: false }],
+      'react/jsx-filename-extension': ['error', { extensions: ['.tsx'] }],
+      'import/no-duplicates': 'error',
+      'tsdoc/syntax': 'error',
+      '@typescript-eslint/ban-ts-comment': 'error',
+      '@typescript-eslint/no-unused-vars': 'error',
+      '@typescript-eslint/no-explicit-any': 'error',
+      '@typescript-eslint/no-non-null-assertion': 'error',
+      '@typescript-eslint/consistent-type-assertions': 'error',
+      '@typescript-eslint/naming-convention': [
+        'error',
+        { selector: 'interface', format: ['PascalCase'], prefix: ['I'] },
+        { selector: ['typeAlias', 'typeLike', 'enum'], format: ['PascalCase'] },
+        { selector: 'typeParameter', format: ['PascalCase'], prefix: ['T'] },
+        {
+          selector: 'variable',
+          format: ['camelCase', 'UPPER_CASE', 'PascalCase'],
+          leadingUnderscore: 'allow',
+        },
+        {
+          selector: 'parameter',
+          format: ['camelCase'],
+          leadingUnderscore: 'allow',
+        },
+        { selector: 'function', format: ['camelCase', 'PascalCase'] },
+        {
+          selector: 'memberLike',
+          modifiers: ['private'],
+          format: ['camelCase'],
+          leadingUnderscore: 'require',
+        },
+        { selector: 'variable', modifiers: ['exported'], format: null },
       ],
-      "react/jsx-pascal-case": ["error", { allowAllCaps: false, allowNamespace: false }],
-      "react/no-this-in-sfc": "error",
-      "react/no-unstable-nested-components": ["error", { allowAsProps: true }],
-      "prettier/prettier": "error",
-      "vitest/no-disabled-tests": "warn",
-      "vitest/no-focused-tests": "error",
-      "vitest/no-identical-title": "error",
-      "@typescript-eslint/no-unused-expressions": "error",
+      'react/jsx-pascal-case': [
+        'error',
+        { allowAllCaps: false, allowNamespace: false },
+      ],
+      'react/no-this-in-sfc': 'error',
+      'react/no-unstable-nested-components': ['error', { allowAsProps: true }],
+      'prettier/prettier': 'error',
+      'vitest/no-disabled-tests': 'warn',
+      'vitest/no-focused-tests': 'error',
+      'vitest/no-identical-title': 'error',
+      '@typescript-eslint/no-unused-expressions': 'error',
     },
   },
   {
-    files: ["*.graphql"],
+    files: ['*.graphql'],
     languageOptions: {
       parser: graphql.parser,
     },
     plugins: {
-      "@graphql-eslint": graphql,
+      '@graphql-eslint': graphql,
       prettier,
     },
     rules: {
-      "@typescript-eslint/consistent-type-imports": "off",
-      "@typescript-eslint/naming-convention": "off",
-      "@typescript-eslint/explicit-function-return-type": "off",
-      "@graphql-eslint/known-type-names": "error",
-      "@graphql-eslint/no-unreachable-types": "off",
-      "prettier/prettier": ["error", { endOfLine: "auto" }],
+      '@typescript-eslint/consistent-type-imports': 'off',
+      '@typescript-eslint/naming-convention': 'off',
+      '@typescript-eslint/explicit-function-return-type': 'off',
+      '@graphql-eslint/known-type-names': 'error',
+      '@graphql-eslint/no-unreachable-types': 'off',
+      'prettier/prettier': ['error', { endOfLine: 'auto' }],
     },
   },
 ];
