--- src/components/UserPortal/ChatRoom/mocks.ts
@@ -1,5295 +1,0 @@
-import {
-  CHATS_LIST,
-  CHAT_BY_ID,
-  GROUP_CHAT_LIST,
-  UNREAD_CHAT_LIST,
-} from 'GraphQl/Queries/PlugInQueries';
-import {
-  MARK_CHAT_MESSAGES_AS_READ,
-  MESSAGE_SENT_TO_CHAT,
-  SEND_MESSAGE_TO_CHAT,
-} from 'GraphQl/Mutations/OrganizationMutations';
-
-export const MESSAGE_SENT_TO_CHAT_MOCK = [
-  {
-    request: {
-      query: MESSAGE_SENT_TO_CHAT,
-      variables: {
-        userId: null,
-      },
-    },
-    result: {
-      data: {
-        messageSentToChat: {
-          _id: '668ec1f1364e03ac47a151',
-          createdAt: '2024-07-10T17:16:33.248Z',
-          chatMessageBelongsTo: {
-            _id: '1',
-          },
-          messageContent: 'Test ',
-          media: null,
-          replyTo: null,
-          sender: {
-            _id: '64378abd85008f171cf2990d',
-            firstName: 'Wilt',
-            lastName: 'Shepherd',
-            image: '',
-          },
-          updatedAt: '2024-07-10',
-        },
-      },
-    },
-  },
-  {
-    request: {
-      query: MESSAGE_SENT_TO_CHAT,
-      variables: {
-        userId: '2',
-      },
-    },
-    result: {
-      data: {
-        messageSentToChat: {
-          _id: '668ec1f1df364e03ac47a151',
-          createdAt: '2024-07-10T17:16:33.248Z',
-          messageContent: 'Test ',
-          media: null,
-          chatMessageBelongsTo: {
-            _id: '1',
-          },
-          replyTo: null,
-          sender: {
-            _id: '64378abd85008f171cf2990d',
-            firstName: 'Wilt',
-            lastName: 'Shepherd',
-            image: '',
-          },
-          updatedAt: '2024-07-10',
-        },
-      },
-    },
-  },
-  {
-    request: {
-      query: MESSAGE_SENT_TO_CHAT,
-      variables: {
-        userId: '8',
-      },
-    },
-    result: {
-      data: {
-        messageSentToChat: {
-          _id: '668ec1f1df364e03ac47a151',
-          createdAt: '2024-07-10T17:16:33.248Z',
-          messageContent: 'Test ',
-          media: null,
-          chatMessageBelongsTo: {
-            _id: '1',
-          },
-          replyTo: null,
-          sender: {
-            _id: '64378abd85008f171cf2990d',
-            firstName: 'Wilt',
-            lastName: 'Shepherd',
-            image: '',
-          },
-          updatedAt: '2024-07-10',
-        },
-      },
-    },
-  },
-  {
-    request: {
-      query: MESSAGE_SENT_TO_CHAT,
-      variables: {
-        userId: '1',
-      },
-    },
-    result: {
-      data: {
-        messageSentToChat: {
-          _id: '668ec1f13603ac4697a151',
-          createdAt: '2024-07-10T17:16:33.248Z',
-          messageContent: 'Test ',
-          media: null,
-          chatMessageBelongsTo: {
-            _id: '1',
-          },
-          replyTo: null,
-          sender: {
-            _id: '64378abd85008f171cf2990d',
-            firstName: 'Wilt',
-            lastName: 'Shepherd',
-            image: '',
-          },
-          updatedAt: '2024-07-10',
-        },
-      },
-    },
-  },
-];
-
-export const UNREAD_CHAT_LIST_QUERY_MOCK = [
-  {
-    request: {
-      query: UNREAD_CHAT_LIST,
-      variables: {},
-    },
-    result: {
-      data: {
-        getUnreadChatsByUserId: {
-          _id: '1',
-          createdAt: '2345678903456',
-          isGroup: false,
-          creator: {
-            _id: '64378abd85008f171cf2990d',
-            firstName: 'Wilt',
-            lastName: 'Shepherd',
-            image: null,
-            email: 'testsuperadmin@example.com',
-            createdAt: '2023-04-13T04:53:17.742Z',
-            __typename: 'User',
-          },
-          organization: null,
-          name: '',
-          messages: [
-            {
-              _id: '345678',
-              createdAt: '345678908765',
-              messageContent: 'Hello',
-              media: null,
-              replyTo: null,
-              type: 'STRING',
-              sender: {
-                _id: '2',
-                firstName: 'Test',
-                lastName: 'User',
-                email: 'test@example.com',
-                image: '',
-              },
-            },
-          ],
-          users: [
-            {
-              _id: '1',
-              firstName: 'Disha',
-              lastName: 'Talreja',
-              email: 'disha@example.com',
-              image: '',
-            },
-            {
-              _id: '2',
-              firstName: 'Test',
-              lastName: 'User',
-              email: 'test@example.com',
-              image: '',
-            },
-          ],
-          admins: [],
-          unseenMessagesByUsers: JSON.stringify({
-            '1': 0,
-            '2': 0,
-          }),
-        },
-      },
-    },
-  },
-  {
-    request: {
-      query: UNREAD_CHAT_LIST,
-      variables: {
-        id: '8',
-      },
-    },
-    result: {
-      data: {
-        getUnreadChatsByUserId: {
-          _id: '1',
-          createdAt: '2345678903456',
-          isGroup: false,
-          creator: {
-            _id: '64378abd85008f171cf2990d',
-            firstName: 'Wilt',
-            lastName: 'Shepherd',
-            image: null,
-            email: 'testsuperadmin@example.com',
-            createdAt: '2023-04-13T04:53:17.742Z',
-            __typename: 'User',
-          },
-          organization: null,
-          name: '',
-          messages: [
-            {
-              _id: '345678',
-              createdAt: '345678908765',
-              messageContent: 'Hello',
-              media: null,
-              replyTo: null,
-              type: 'STRING',
-              sender: {
-                _id: '2',
-                firstName: 'Test',
-                lastName: 'User',
-                email: 'test@example.com',
-                image: '',
-              },
-            },
-          ],
-          users: [
-            {
-              _id: '1',
-              firstName: 'Disha',
-              lastName: 'Talreja',
-              email: 'disha@example.com',
-              image: '',
-            },
-            {
-              _id: '2',
-              firstName: 'Test',
-              lastName: 'User',
-              email: 'test@example.com',
-              image: '',
-            },
-          ],
-          admins: [],
-          unseenMessagesByUsers: JSON.stringify({
-            '1': 0,
-            '2': 0,
-          }),
-        },
-      },
-    },
-  },
-  {
-    request: {
-      query: UNREAD_CHAT_LIST,
-      variables: {
-        id: '8',
-      },
-    },
-    result: {
-      data: {
-        getUnreadChatsByUserId: {
-          _id: '1',
-          createdAt: '2345678903456',
-          isGroup: false,
-          creator: {
-            _id: '64378abd85008f171cf2990d',
-            firstName: 'Wilt',
-            lastName: 'Shepherd',
-            image: null,
-            email: 'testsuperadmin@example.com',
-            createdAt: '2023-04-13T04:53:17.742Z',
-            __typename: 'User',
-          },
-          organization: null,
-          name: '',
-          messages: [
-            {
-              _id: '345678',
-              createdAt: '345678908765',
-              messageContent: 'Hello',
-              media: null,
-              replyTo: null,
-              type: 'STRING',
-              sender: {
-                _id: '2',
-                firstName: 'Test',
-                lastName: 'User',
-                email: 'test@example.com',
-                image: '',
-              },
-            },
-          ],
-          users: [
-            {
-              _id: '1',
-              firstName: 'Disha',
-              lastName: 'Talreja',
-              email: 'disha@example.com',
-              image: '',
-            },
-            {
-              _id: '2',
-              firstName: 'Test',
-              lastName: 'User',
-              email: 'test@example.com',
-              image: '',
-            },
-          ],
-          admins: [],
-          unseenMessagesByUsers: JSON.stringify({
-            '1': 0,
-            '2': 0,
-          }),
-        },
-      },
-    },
-  },
-  {
-    request: {
-      query: UNREAD_CHAT_LIST,
-      variables: {},
-    },
-    result: {
-      data: {
-        getUnreadChatsByUserId: {
-          _id: '1',
-          createdAt: '2345678903456',
-          isGroup: false,
-          creator: {
-            _id: '64378abd85008f171cf2990d',
-            firstName: 'Wilt',
-            lastName: 'Shepherd',
-            image: null,
-            email: 'testsuperadmin@example.com',
-            createdAt: '2023-04-13T04:53:17.742Z',
-            __typename: 'User',
-          },
-          organization: null,
-          name: '',
-          messages: [
-            {
-              _id: '345678',
-              createdAt: '345678908765',
-              messageContent: 'Hello',
-              media: null,
-              replyTo: null,
-              type: 'STRING',
-              sender: {
-                _id: '2',
-                firstName: 'Test',
-                lastName: 'User',
-                email: 'test@example.com',
-                image: '',
-              },
-            },
-          ],
-          users: [
-            {
-              _id: '1',
-              firstName: 'Disha',
-              lastName: 'Talreja',
-              email: 'disha@example.com',
-              image: '',
-            },
-            {
-              _id: '2',
-              firstName: 'Test',
-              lastName: 'User',
-              email: 'test@example.com',
-              image: '',
-            },
-          ],
-          admins: [],
-          unseenMessagesByUsers: JSON.stringify({
-            '1': 0,
-            '2': 0,
-          }),
-        },
-      },
-    },
-  },
-  {
-    request: {
-      query: UNREAD_CHAT_LIST,
-      variables: {},
-    },
-    result: {
-      data: {
-        getUnreadChatsByUserId: {
-          _id: '1',
-          createdAt: '2345678903456',
-          isGroup: false,
-          creator: {
-            _id: '64378abd85008f171cf2990d',
-            firstName: 'Wilt',
-            lastName: 'Shepherd',
-            image: null,
-            email: 'testsuperadmin@example.com',
-            createdAt: '2023-04-13T04:53:17.742Z',
-            __typename: 'User',
-          },
-          organization: null,
-          name: '',
-          messages: [
-            {
-              _id: '345678',
-              createdAt: '345678908765',
-              messageContent: 'Hello',
-              media: null,
-              replyTo: null,
-              type: 'STRING',
-              sender: {
-                _id: '2',
-                firstName: 'Test',
-                lastName: 'User',
-                email: 'test@example.com',
-                image: '',
-              },
-            },
-          ],
-          users: [
-            {
-              _id: '1',
-              firstName: 'Disha',
-              lastName: 'Talreja',
-              email: 'disha@example.com',
-              image: '',
-            },
-            {
-              _id: '2',
-              firstName: 'Test',
-              lastName: 'User',
-              email: 'test@example.com',
-              image: '',
-            },
-          ],
-          admins: [],
-          unseenMessagesByUsers: JSON.stringify({
-            '1': 0,
-            '2': 0,
-          }),
-        },
-      },
-    },
-  },
-  {
-    request: {
-      query: UNREAD_CHAT_LIST,
-      variables: {},
-    },
-    result: {
-      data: {
-        getUnreadChatsByUserId: {
-          _id: '1',
-          createdAt: '2345678903456',
-          isGroup: false,
-          creator: {
-            _id: '64378abd85008f171cf2990d',
-            firstName: 'Wilt',
-            lastName: 'Shepherd',
-            image: null,
-            email: 'testsuperadmin@example.com',
-            createdAt: '2023-04-13T04:53:17.742Z',
-            __typename: 'User',
-          },
-          organization: null,
-          name: '',
-          messages: [
-            {
-              _id: '345678',
-              createdAt: '345678908765',
-              messageContent: 'Hello',
-              media: null,
-              replyTo: null,
-              type: 'STRING',
-              sender: {
-                _id: '2',
-                firstName: 'Test',
-                lastName: 'User',
-                email: 'test@example.com',
-                image: '',
-              },
-            },
-          ],
-          users: [
-            {
-              _id: '1',
-              firstName: 'Disha',
-              lastName: 'Talreja',
-              email: 'disha@example.com',
-              image: '',
-            },
-            {
-              _id: '2',
-              firstName: 'Test',
-              lastName: 'User',
-              email: 'test@example.com',
-              image: '',
-            },
-          ],
-          admins: [],
-          unseenMessagesByUsers: JSON.stringify({
-            '1': 0,
-            '2': 0,
-          }),
-        },
-      },
-    },
-  },
-  {
-    request: {
-      query: UNREAD_CHAT_LIST,
-      variables: {
-        id: '1',
-      },
-    },
-    result: {
-      data: {
-        getUnreadChatsByUserId: {
-          _id: '1',
-          createdAt: '2345678903456',
-          isGroup: false,
-          creator: {
-            _id: '64378abd85008f171cf2990d',
-            firstName: 'Wilt',
-            lastName: 'Shepherd',
-            image: null,
-            email: 'testsuperadmin@example.com',
-            createdAt: '2023-04-13T04:53:17.742Z',
-            __typename: 'User',
-          },
-          organization: null,
-          name: '',
-          messages: [
-            {
-              _id: '345678',
-              createdAt: '345678908765',
-              messageContent: 'Hello',
-              media: null,
-              replyTo: null,
-              type: 'STRING',
-              sender: {
-                _id: '2',
-                firstName: 'Test',
-                lastName: 'User',
-                email: 'test@example.com',
-                image: '',
-              },
-            },
-          ],
-          users: [
-            {
-              _id: '1',
-              firstName: 'Disha',
-              lastName: 'Talreja',
-              email: 'disha@example.com',
-              image: '',
-            },
-            {
-              _id: '2',
-              firstName: 'Test',
-              lastName: 'User',
-              email: 'test@example.com',
-              image: '',
-            },
-          ],
-          admins: [],
-          unseenMessagesByUsers: JSON.stringify({
-            '1': 0,
-            '2': 0,
-          }),
-        },
-      },
-    },
-  },
-  {
-    request: {
-      query: UNREAD_CHAT_LIST,
-      variables: {
-        id: '1',
-      },
-    },
-    result: {
-      data: {
-        getUnreadChatsByUserId: {
-          _id: '1',
-          createdAt: '2345678903456',
-          isGroup: false,
-          creator: {
-            _id: '64378abd85008f171cf2990d',
-            firstName: 'Wilt',
-            lastName: 'Shepherd',
-            image: null,
-            email: 'testsuperadmin@example.com',
-            createdAt: '2023-04-13T04:53:17.742Z',
-            __typename: 'User',
-          },
-          organization: null,
-          name: '',
-          messages: [
-            {
-              _id: '345678',
-              createdAt: '345678908765',
-              messageContent: 'Hello',
-              media: null,
-              replyTo: null,
-              type: 'STRING',
-              sender: {
-                _id: '2',
-                firstName: 'Test',
-                lastName: 'User',
-                email: 'test@example.com',
-                image: '',
-              },
-            },
-          ],
-          users: [
-            {
-              _id: '1',
-              firstName: 'Disha',
-              lastName: 'Talreja',
-              email: 'disha@example.com',
-              image: '',
-            },
-            {
-              _id: '2',
-              firstName: 'Test',
-              lastName: 'User',
-              email: 'test@example.com',
-              image: '',
-            },
-          ],
-          admins: [],
-          unseenMessagesByUsers: JSON.stringify({
-            '1': 0,
-            '2': 0,
-          }),
-        },
-      },
-    },
-  },
-  {
-    request: {
-      query: UNREAD_CHAT_LIST,
-      variables: {
-        id: '1',
-      },
-    },
-    result: {
-      data: {
-        getUnreadChatsByUserId: {
-          _id: '1',
-          createdAt: '2345678903456',
-          isGroup: false,
-          creator: {
-            _id: '64378abd85008f171cf2990d',
-            firstName: 'Wilt',
-            lastName: 'Shepherd',
-            image: null,
-            email: 'testsuperadmin@example.com',
-            createdAt: '2023-04-13T04:53:17.742Z',
-            __typename: 'User',
-          },
-          organization: null,
-          name: '',
-          messages: [
-            {
-              _id: '345678',
-              createdAt: '345678908765',
-              messageContent: 'Hello',
-              media: null,
-              replyTo: null,
-              type: 'STRING',
-              sender: {
-                _id: '2',
-                firstName: 'Test',
-                lastName: 'User',
-                email: 'test@example.com',
-                image: '',
-              },
-            },
-          ],
-          users: [
-            {
-              _id: '1',
-              firstName: 'Disha',
-              lastName: 'Talreja',
-              email: 'disha@example.com',
-              image: '',
-            },
-            {
-              _id: '2',
-              firstName: 'Test',
-              lastName: 'User',
-              email: 'test@example.com',
-              image: '',
-            },
-          ],
-          admins: [],
-          unseenMessagesByUsers: JSON.stringify({
-            '1': 0,
-            '2': 0,
-          }),
-        },
-      },
-    },
-  },
-  {
-    request: {
-      query: UNREAD_CHAT_LIST,
-      variables: {
-        id: null,
-      },
-    },
-    result: {
-      data: {
-        getUnreadChatsByUserId: {
-          _id: '1',
-          createdAt: '2345678903456',
-          isGroup: false,
-          creator: {
-            _id: '64378abd85008f171cf2990d',
-            firstName: 'Wilt',
-            lastName: 'Shepherd',
-            image: null,
-            email: 'testsuperadmin@example.com',
-            createdAt: '2023-04-13T04:53:17.742Z',
-            __typename: 'User',
-          },
-          organization: null,
-          name: '',
-          messages: [
-            {
-              _id: '345678',
-              createdAt: '345678908765',
-              messageContent: 'Hello',
-              media: null,
-              replyTo: null,
-              type: 'STRING',
-              sender: {
-                _id: '2',
-                firstName: 'Test',
-                lastName: 'User',
-                email: 'test@example.com',
-                image: '',
-              },
-            },
-          ],
-          users: [
-            {
-              _id: '1',
-              firstName: 'Disha',
-              lastName: 'Talreja',
-              email: 'disha@example.com',
-              image: '',
-            },
-            {
-              _id: '2',
-              firstName: 'Test',
-              lastName: 'User',
-              email: 'test@example.com',
-              image: '',
-            },
-          ],
-          admins: [],
-          unseenMessagesByUsers: JSON.stringify({
-            '1': 0,
-            '2': 0,
-          }),
-        },
-      },
-    },
-  },
-  {
-    request: {
-      query: UNREAD_CHAT_LIST,
-      variables: {
-        id: null,
-      },
-    },
-    result: {
-      data: {
-        getUnreadChatsByUserId: {
-          _id: '1',
-          createdAt: '2345678903456',
-          isGroup: false,
-          creator: {
-            _id: '64378abd85008f171cf2990d',
-            firstName: 'Wilt',
-            lastName: 'Shepherd',
-            image: null,
-            email: 'testsuperadmin@example.com',
-            createdAt: '2023-04-13T04:53:17.742Z',
-            __typename: 'User',
-          },
-          organization: null,
-          name: '',
-          messages: [
-            {
-              _id: '345678',
-              createdAt: '345678908765',
-              messageContent: 'Hello',
-              media: null,
-              replyTo: null,
-              type: 'STRING',
-              sender: {
-                _id: '2',
-                firstName: 'Test',
-                lastName: 'User',
-                email: 'test@example.com',
-                image: '',
-              },
-            },
-          ],
-          users: [
-            {
-              _id: '1',
-              firstName: 'Disha',
-              lastName: 'Talreja',
-              email: 'disha@example.com',
-              image: '',
-            },
-            {
-              _id: '2',
-              firstName: 'Test',
-              lastName: 'User',
-              email: 'test@example.com',
-              image: '',
-            },
-          ],
-          admins: [],
-          unseenMessagesByUsers: JSON.stringify({
-            '1': 0,
-            '2': 0,
-          }),
-        },
-      },
-    },
-  },
-  {
-    request: {
-      query: UNREAD_CHAT_LIST,
-      variables: {
-        id: '8',
-      },
-    },
-    result: {
-      data: {
-        getUnreadChatsByUserId: {
-          _id: '1',
-          createdAt: '2345678903456',
-          isGroup: false,
-          creator: {
-            _id: '64378abd85008f171cf2990d',
-            firstName: 'Wilt',
-            lastName: 'Shepherd',
-            image: null,
-            email: 'testsuperadmin@example.com',
-            createdAt: '2023-04-13T04:53:17.742Z',
-            __typename: 'User',
-          },
-          organization: null,
-          name: '',
-          messages: [
-            {
-              _id: '345678',
-              createdAt: '345678908765',
-              messageContent: 'Hello',
-              media: null,
-              replyTo: null,
-              type: 'STRING',
-              sender: {
-                _id: '2',
-                firstName: 'Test',
-                lastName: 'User',
-                email: 'test@example.com',
-                image: '',
-              },
-            },
-          ],
-          users: [
-            {
-              _id: '1',
-              firstName: 'Disha',
-              lastName: 'Talreja',
-              email: 'disha@example.com',
-              image: '',
-            },
-            {
-              _id: '2',
-              firstName: 'Test',
-              lastName: 'User',
-              email: 'test@example.com',
-              image: '',
-            },
-          ],
-          admins: [],
-          unseenMessagesByUsers: JSON.stringify({
-            '1': 0,
-            '2': 0,
-          }),
-        },
-      },
-    },
-  },
-  {
-    request: {
-      query: UNREAD_CHAT_LIST,
-      variables: {
-        id: '8',
-      },
-    },
-    result: {
-      data: {
-        getUnreadChatsByUserId: {
-          _id: '1',
-          createdAt: '2345678903456',
-          isGroup: false,
-          creator: {
-            _id: '64378abd85008f171cf2990d',
-            firstName: 'Wilt',
-            lastName: 'Shepherd',
-            image: null,
-            email: 'testsuperadmin@example.com',
-            createdAt: '2023-04-13T04:53:17.742Z',
-            __typename: 'User',
-          },
-          organization: null,
-          name: '',
-          messages: [
-            {
-              _id: '345678',
-              createdAt: '345678908765',
-              messageContent: 'Hello',
-              media: null,
-              replyTo: null,
-              type: 'STRING',
-              sender: {
-                _id: '2',
-                firstName: 'Test',
-                lastName: 'User',
-                email: 'test@example.com',
-                image: '',
-              },
-            },
-          ],
-          users: [
-            {
-              _id: '1',
-              firstName: 'Disha',
-              lastName: 'Talreja',
-              email: 'disha@example.com',
-              image: '',
-            },
-            {
-              _id: '2',
-              firstName: 'Test',
-              lastName: 'User',
-              email: 'test@example.com',
-              image: '',
-            },
-          ],
-          admins: [],
-          unseenMessagesByUsers: JSON.stringify({
-            '1': 0,
-            '2': 0,
-          }),
-        },
-      },
-    },
-  },
-  {
-    request: {
-      query: UNREAD_CHAT_LIST,
-      variables: {
-        id: '8',
-      },
-    },
-    result: {
-      data: {
-        getUnreadChatsByUserId: {
-          _id: '1',
-          createdAt: '2345678903456',
-          isGroup: false,
-          creator: {
-            _id: '64378abd85008f171cf2990d',
-            firstName: 'Wilt',
-            lastName: 'Shepherd',
-            image: null,
-            email: 'testsuperadmin@example.com',
-            createdAt: '2023-04-13T04:53:17.742Z',
-            __typename: 'User',
-          },
-          organization: null,
-          name: '',
-          messages: [
-            {
-              _id: '345678',
-              createdAt: '345678908765',
-              messageContent: 'Hello',
-              media: null,
-              replyTo: null,
-              type: 'STRING',
-              sender: {
-                _id: '2',
-                firstName: 'Test',
-                lastName: 'User',
-                email: 'test@example.com',
-                image: '',
-              },
-            },
-          ],
-          users: [
-            {
-              _id: '1',
-              firstName: 'Disha',
-              lastName: 'Talreja',
-              email: 'disha@example.com',
-              image: '',
-            },
-            {
-              _id: '2',
-              firstName: 'Test',
-              lastName: 'User',
-              email: 'test@example.com',
-              image: '',
-            },
-          ],
-          admins: [],
-          unseenMessagesByUsers: JSON.stringify({
-            '1': 0,
-            '2': 0,
-          }),
-        },
-      },
-    },
-  },
-];
-
-export const GROUP_CHAT_LIST_QUERY_MOCK = [
-  {
-    request: {
-      query: GROUP_CHAT_LIST,
-      variables: {
-        id: '',
-      },
-    },
-    result: {
-      data: {
-        getGroupChatsByUserId: {
-          _id: '65844efc814dd4003db811c4',
-          isGroup: true,
-          creator: {
-            _id: '64378abd85008f171cf2990d',
-            firstName: 'Wilt',
-            lastName: 'Shepherd',
-            image: null,
-            email: 'testsuperadmin@example.com',
-            createdAt: '2023-04-13T04:53:17.742Z',
-            __typename: 'User',
-          },
-          organization: {
-            _id: 'pw3ertyuiophgfre45678',
-            name: 'rtyu',
-          },
-          createdAt: '2345678903456',
-          name: 'Test Group Chat',
-          messages: [
-            {
-              _id: '345678',
-              createdAt: '345678908765',
-              messageContent: 'Hello',
-              media: null,
-              replyTo: null,
-              type: 'STRING',
-              sender: {
-                _id: '2',
-                firstName: 'Test',
-                lastName: 'User',
-                email: 'test@example.com',
-                image: '',
-              },
-            },
-          ],
-          users: [
-            {
-              _id: '1',
-              firstName: 'Disha',
-              lastName: 'Talreja',
-              email: 'disha@example.com',
-              image: '',
-            },
-            {
-              _id: '2',
-              firstName: 'Test',
-              lastName: 'User',
-              email: 'test@example.com',
-              image: '',
-            },
-            {
-              _id: '3',
-              firstName: 'Test',
-              lastName: 'User1',
-              email: 'test1@example.com',
-              image: '',
-            },
-            {
-              _id: '4',
-              firstName: 'Test',
-              lastName: 'User2',
-              email: 'test2@example.com',
-              image: '',
-            },
-            {
-              _id: '5',
-              firstName: 'Test',
-              lastName: 'User4',
-              email: 'test4@example.com',
-              image: '',
-            },
-          ],
-          admins: [
-            {
-              _id: '1',
-              firstName: 'Disha',
-              lastName: 'Talreja',
-              email: 'disha@example.com',
-              image: '',
-            },
-          ],
-          unseenMessagesByUsers: JSON.stringify({
-            '1': 1,
-            '2': 1,
-            '3': 1,
-            '4': 1,
-            '5': 1,
-          }),
-        },
-      },
-    },
-  },
-  {
-    request: {
-      query: GROUP_CHAT_LIST,
-      variables: {},
-    },
-    result: {
-      data: {
-        getGroupChatsByUserId: {
-          _id: '65844efc814dd4003db811c4',
-          isGroup: true,
-          creator: {
-            _id: '64378abd85008f171cf2990d',
-            firstName: 'Wilt',
-            lastName: 'Shepherd',
-            image: null,
-            email: 'testsuperadmin@example.com',
-            createdAt: '2023-04-13T04:53:17.742Z',
-            __typename: 'User',
-          },
-          organization: {
-            _id: 'pw3ertyuiophgfre45678',
-            name: 'rtyu',
-          },
-          createdAt: '2345678903456',
-          name: 'Test Group Chat',
-          messages: [
-            {
-              _id: '345678',
-              createdAt: '345678908765',
-              messageContent: 'Hello',
-              media: null,
-              replyTo: null,
-              type: 'STRING',
-              sender: {
-                _id: '2',
-                firstName: 'Test',
-                lastName: 'User',
-                email: 'test@example.com',
-                image: '',
-              },
-            },
-          ],
-          users: [
-            {
-              _id: '1',
-              firstName: 'Disha',
-              lastName: 'Talreja',
-              email: 'disha@example.com',
-              image: '',
-            },
-            {
-              _id: '2',
-              firstName: 'Test',
-              lastName: 'User',
-              email: 'test@example.com',
-              image: '',
-            },
-            {
-              _id: '3',
-              firstName: 'Test',
-              lastName: 'User1',
-              email: 'test1@example.com',
-              image: '',
-            },
-            {
-              _id: '4',
-              firstName: 'Test',
-              lastName: 'User2',
-              email: 'test2@example.com',
-              image: '',
-            },
-            {
-              _id: '5',
-              firstName: 'Test',
-              lastName: 'User4',
-              email: 'test4@example.com',
-              image: '',
-            },
-          ],
-          admins: [
-            {
-              _id: '1',
-              firstName: 'Disha',
-              lastName: 'Talreja',
-              email: 'disha@example.com',
-              image: '',
-            },
-          ],
-          unseenMessagesByUsers: JSON.stringify({
-            '1': 1,
-            '2': 1,
-            '3': 1,
-            '4': 1,
-            '5': 1,
-          }),
-        },
-      },
-    },
-  },
-  {
-    request: {
-      query: GROUP_CHAT_LIST,
-      variables: {},
-    },
-    result: {
-      data: {
-        getGroupChatsByUserId: {
-          _id: '65844efc814dd4003db811c4',
-          isGroup: true,
-          creator: {
-            _id: '64378abd85008f171cf2990d',
-            firstName: 'Wilt',
-            lastName: 'Shepherd',
-            image: null,
-            email: 'testsuperadmin@example.com',
-            createdAt: '2023-04-13T04:53:17.742Z',
-            __typename: 'User',
-          },
-          organization: {
-            _id: 'pw3ertyuiophgfre45678',
-            name: 'rtyu',
-          },
-          createdAt: '2345678903456',
-          name: 'Test Group Chat',
-          messages: [
-            {
-              _id: '345678',
-              createdAt: '345678908765',
-              messageContent: 'Hello',
-              media: null,
-              replyTo: null,
-              type: 'STRING',
-              sender: {
-                _id: '2',
-                firstName: 'Test',
-                lastName: 'User',
-                email: 'test@example.com',
-                image: '',
-              },
-            },
-          ],
-          users: [
-            {
-              _id: '1',
-              firstName: 'Disha',
-              lastName: 'Talreja',
-              email: 'disha@example.com',
-              image: '',
-            },
-            {
-              _id: '2',
-              firstName: 'Test',
-              lastName: 'User',
-              email: 'test@example.com',
-              image: '',
-            },
-            {
-              _id: '3',
-              firstName: 'Test',
-              lastName: 'User1',
-              email: 'test1@example.com',
-              image: '',
-            },
-            {
-              _id: '4',
-              firstName: 'Test',
-              lastName: 'User2',
-              email: 'test2@example.com',
-              image: '',
-            },
-            {
-              _id: '5',
-              firstName: 'Test',
-              lastName: 'User4',
-              email: 'test4@example.com',
-              image: '',
-            },
-          ],
-          admins: [
-            {
-              _id: '1',
-              firstName: 'Disha',
-              lastName: 'Talreja',
-              email: 'disha@example.com',
-              image: '',
-            },
-          ],
-          unseenMessagesByUsers: JSON.stringify({
-            '1': 1,
-            '2': 1,
-            '3': 1,
-            '4': 1,
-            '5': 1,
-          }),
-        },
-      },
-    },
-  },
-  {
-    request: {
-      query: GROUP_CHAT_LIST,
-      variables: {},
-    },
-    result: {
-      data: {
-        getGroupChatsByUserId: {
-          _id: '65844efc814dd4003db811c4',
-          isGroup: true,
-          creator: {
-            _id: '64378abd85008f171cf2990d',
-            firstName: 'Wilt',
-            lastName: 'Shepherd',
-            image: null,
-            email: 'testsuperadmin@example.com',
-            createdAt: '2023-04-13T04:53:17.742Z',
-            __typename: 'User',
-          },
-          organization: {
-            _id: 'pw3ertyuiophgfre45678',
-            name: 'rtyu',
-          },
-          createdAt: '2345678903456',
-          name: 'Test Group Chat',
-          messages: [
-            {
-              _id: '345678',
-              createdAt: '345678908765',
-              messageContent: 'Hello',
-              media: null,
-              replyTo: null,
-              type: 'STRING',
-              sender: {
-                _id: '2',
-                firstName: 'Test',
-                lastName: 'User',
-                email: 'test@example.com',
-                image: '',
-              },
-            },
-          ],
-          users: [
-            {
-              _id: '1',
-              firstName: 'Disha',
-              lastName: 'Talreja',
-              email: 'disha@example.com',
-              image: '',
-            },
-            {
-              _id: '2',
-              firstName: 'Test',
-              lastName: 'User',
-              email: 'test@example.com',
-              image: '',
-            },
-            {
-              _id: '3',
-              firstName: 'Test',
-              lastName: 'User1',
-              email: 'test1@example.com',
-              image: '',
-            },
-            {
-              _id: '4',
-              firstName: 'Test',
-              lastName: 'User2',
-              email: 'test2@example.com',
-              image: '',
-            },
-            {
-              _id: '5',
-              firstName: 'Test',
-              lastName: 'User4',
-              email: 'test4@example.com',
-              image: '',
-            },
-          ],
-          admins: [
-            {
-              _id: '1',
-              firstName: 'Disha',
-              lastName: 'Talreja',
-              email: 'disha@example.com',
-              image: '',
-            },
-          ],
-          unseenMessagesByUsers: JSON.stringify({
-            '1': 1,
-            '2': 1,
-            '3': 1,
-            '4': 1,
-            '5': 1,
-          }),
-        },
-      },
-    },
-  },
-  {
-    request: {
-      query: GROUP_CHAT_LIST,
-      variables: {},
-    },
-    result: {
-      data: {
-        getGroupChatsByUserId: {
-          _id: '65844efc814dd4003db811c4',
-          isGroup: true,
-          creator: {
-            _id: '64378abd85008f171cf2990d',
-            firstName: 'Wilt',
-            lastName: 'Shepherd',
-            image: null,
-            email: 'testsuperadmin@example.com',
-            createdAt: '2023-04-13T04:53:17.742Z',
-            __typename: 'User',
-          },
-          organization: {
-            _id: 'pw3ertyuiophgfre45678',
-            name: 'rtyu',
-          },
-          createdAt: '2345678903456',
-          name: 'Test Group Chat',
-          messages: [
-            {
-              _id: '345678',
-              createdAt: '345678908765',
-              messageContent: 'Hello',
-              media: null,
-              replyTo: null,
-              type: 'STRING',
-              sender: {
-                _id: '2',
-                firstName: 'Test',
-                lastName: 'User',
-                email: 'test@example.com',
-                image: '',
-              },
-            },
-          ],
-          users: [
-            {
-              _id: '1',
-              firstName: 'Disha',
-              lastName: 'Talreja',
-              email: 'disha@example.com',
-              image: '',
-            },
-            {
-              _id: '2',
-              firstName: 'Test',
-              lastName: 'User',
-              email: 'test@example.com',
-              image: '',
-            },
-            {
-              _id: '3',
-              firstName: 'Test',
-              lastName: 'User1',
-              email: 'test1@example.com',
-              image: '',
-            },
-            {
-              _id: '4',
-              firstName: 'Test',
-              lastName: 'User2',
-              email: 'test2@example.com',
-              image: '',
-            },
-            {
-              _id: '5',
-              firstName: 'Test',
-              lastName: 'User4',
-              email: 'test4@example.com',
-              image: '',
-            },
-          ],
-          admins: [
-            {
-              _id: '1',
-              firstName: 'Disha',
-              lastName: 'Talreja',
-              email: 'disha@example.com',
-              image: '',
-            },
-          ],
-          unseenMessagesByUsers: JSON.stringify({
-            '1': 1,
-            '2': 1,
-            '3': 1,
-            '4': 1,
-            '5': 1,
-          }),
-        },
-      },
-    },
-  },
-  {
-    request: {
-      query: GROUP_CHAT_LIST,
-      variables: {
-        id: null,
-      },
-    },
-    result: {
-      data: {
-        getGroupChatsByUserId: {
-          _id: '65844efc814dd4003db811c4',
-          isGroup: true,
-          creator: {
-            _id: '64378abd85008f171cf2990d',
-            firstName: 'Wilt',
-            lastName: 'Shepherd',
-            image: null,
-            email: 'testsuperadmin@example.com',
-            createdAt: '2023-04-13T04:53:17.742Z',
-            __typename: 'User',
-          },
-          organization: {
-            _id: 'pw3ertyuiophgfre45678',
-            name: 'rtyu',
-          },
-          createdAt: '2345678903456',
-          name: 'Test Group Chat',
-          messages: [
-            {
-              _id: '345678',
-              createdAt: '345678908765',
-              messageContent: 'Hello',
-              media: null,
-              replyTo: null,
-              type: 'STRING',
-              sender: {
-                _id: '2',
-                firstName: 'Test',
-                lastName: 'User',
-                email: 'test@example.com',
-                image: '',
-              },
-            },
-          ],
-          users: [
-            {
-              _id: '1',
-              firstName: 'Disha',
-              lastName: 'Talreja',
-              email: 'disha@example.com',
-              image: '',
-            },
-            {
-              _id: '2',
-              firstName: 'Test',
-              lastName: 'User',
-              email: 'test@example.com',
-              image: '',
-            },
-            {
-              _id: '3',
-              firstName: 'Test',
-              lastName: 'User1',
-              email: 'test1@example.com',
-              image: '',
-            },
-            {
-              _id: '4',
-              firstName: 'Test',
-              lastName: 'User2',
-              email: 'test2@example.com',
-              image: '',
-            },
-            {
-              _id: '5',
-              firstName: 'Test',
-              lastName: 'User4',
-              email: 'test4@example.com',
-              image: '',
-            },
-          ],
-          admins: [
-            {
-              _id: '1',
-              firstName: 'Disha',
-              lastName: 'Talreja',
-              email: 'disha@example.com',
-              image: '',
-            },
-          ],
-          unseenMessagesByUsers: JSON.stringify({
-            '1': 1,
-            '2': 1,
-            '3': 1,
-            '4': 1,
-            '5': 1,
-          }),
-        },
-      },
-    },
-  },
-  {
-    request: {
-      query: GROUP_CHAT_LIST,
-      variables: {
-        id: null,
-      },
-    },
-    result: {
-      data: {
-        getGroupChatsByUserId: {
-          _id: '65844efc814dd4003db811c4',
-          isGroup: true,
-          creator: {
-            _id: '64378abd85008f171cf2990d',
-            firstName: 'Wilt',
-            lastName: 'Shepherd',
-            image: null,
-            email: 'testsuperadmin@example.com',
-            createdAt: '2023-04-13T04:53:17.742Z',
-            __typename: 'User',
-          },
-          organization: {
-            _id: 'pw3ertyuiophgfre45678',
-            name: 'rtyu',
-          },
-          createdAt: '2345678903456',
-          name: 'Test Group Chat',
-          messages: [
-            {
-              _id: '345678',
-              createdAt: '345678908765',
-              messageContent: 'Hello',
-              media: null,
-              replyTo: null,
-              type: 'STRING',
-              sender: {
-                _id: '2',
-                firstName: 'Test',
-                lastName: 'User',
-                email: 'test@example.com',
-                image: '',
-              },
-            },
-          ],
-          users: [
-            {
-              _id: '1',
-              firstName: 'Disha',
-              lastName: 'Talreja',
-              email: 'disha@example.com',
-              image: '',
-            },
-            {
-              _id: '2',
-              firstName: 'Test',
-              lastName: 'User',
-              email: 'test@example.com',
-              image: '',
-            },
-            {
-              _id: '3',
-              firstName: 'Test',
-              lastName: 'User1',
-              email: 'test1@example.com',
-              image: '',
-            },
-            {
-              _id: '4',
-              firstName: 'Test',
-              lastName: 'User2',
-              email: 'test2@example.com',
-              image: '',
-            },
-            {
-              _id: '5',
-              firstName: 'Test',
-              lastName: 'User4',
-              email: 'test4@example.com',
-              image: '',
-            },
-          ],
-          admins: [
-            {
-              _id: '1',
-              firstName: 'Disha',
-              lastName: 'Talreja',
-              email: 'disha@example.com',
-              image: '',
-            },
-          ],
-          unseenMessagesByUsers: JSON.stringify({
-            '1': 1,
-            '2': 1,
-            '3': 1,
-            '4': 1,
-            '5': 1,
-          }),
-        },
-      },
-    },
-  },
-  {
-    request: {
-      query: GROUP_CHAT_LIST,
-      variables: {
-        id: null,
-      },
-    },
-    result: {
-      data: {
-        getGroupChatsByUserId: {
-          _id: '65844efc814dd4003db811c4',
-          isGroup: true,
-          creator: {
-            _id: '64378abd85008f171cf2990d',
-            firstName: 'Wilt',
-            lastName: 'Shepherd',
-            image: null,
-            email: 'testsuperadmin@example.com',
-            createdAt: '2023-04-13T04:53:17.742Z',
-            __typename: 'User',
-          },
-          organization: {
-            _id: 'pw3ertyuiophgfre45678',
-            name: 'rtyu',
-          },
-          createdAt: '2345678903456',
-          name: 'Test Group Chat',
-          messages: [
-            {
-              _id: '345678',
-              createdAt: '345678908765',
-              messageContent: 'Hello',
-              media: null,
-              replyTo: null,
-              type: 'STRING',
-              sender: {
-                _id: '2',
-                firstName: 'Test',
-                lastName: 'User',
-                email: 'test@example.com',
-                image: '',
-              },
-            },
-          ],
-          users: [
-            {
-              _id: '1',
-              firstName: 'Disha',
-              lastName: 'Talreja',
-              email: 'disha@example.com',
-              image: '',
-            },
-            {
-              _id: '2',
-              firstName: 'Test',
-              lastName: 'User',
-              email: 'test@example.com',
-              image: '',
-            },
-            {
-              _id: '3',
-              firstName: 'Test',
-              lastName: 'User1',
-              email: 'test1@example.com',
-              image: '',
-            },
-            {
-              _id: '4',
-              firstName: 'Test',
-              lastName: 'User2',
-              email: 'test2@example.com',
-              image: '',
-            },
-            {
-              _id: '5',
-              firstName: 'Test',
-              lastName: 'User4',
-              email: 'test4@example.com',
-              image: '',
-            },
-          ],
-          admins: [
-            {
-              _id: '1',
-              firstName: 'Disha',
-              lastName: 'Talreja',
-              email: 'disha@example.com',
-              image: '',
-            },
-          ],
-          unseenMessagesByUsers: JSON.stringify({
-            '1': 1,
-            '2': 1,
-            '3': 1,
-            '4': 1,
-            '5': 1,
-          }),
-        },
-      },
-    },
-  },
-];
-
-export const CHAT_BY_ID_QUERY_MOCK = [
-  {
-    request: {
-      query: CHAT_BY_ID,
-      variables: {
-        id: '1',
-      },
-    },
-    result: {
-      data: {
-        chatById: {
-          _id: '1',
-          createdAt: '2345678903456',
-          isGroup: false,
-          creator: {
-            _id: '64378abd85008f171cf2990d',
-            firstName: 'Wilt',
-            lastName: 'Shepherd',
-            image: null,
-            email: 'testsuperadmin@example.com',
-            createdAt: '2023-04-13T04:53:17.742Z',
-            __typename: 'User',
-          },
-          organization: null,
-          name: '',
-          messages: [
-            {
-              _id: '4',
-              createdAt: '345678908765',
-              messageContent: 'Hello',
-              media: null,
-              replyTo: null,
-              sender: {
-                _id: '2',
-                firstName: 'Test',
-                lastName: 'User',
-                email: 'test@example.com',
-                image: '',
-              },
-            },
-          ],
-          users: [
-            {
-              _id: '1',
-              firstName: 'Disha',
-              lastName: 'Talreja',
-              email: 'disha@example.com',
-              image: '',
-            },
-            {
-              _id: '2',
-              firstName: 'Test',
-              lastName: 'User',
-              email: 'test@example.com',
-              image: '',
-            },
-          ],
-          admins: [],
-          unseenMessagesByUsers: {
-            '1': 0,
-            '2': 0,
-          },
-        },
-      },
-    },
-  },
-  {
-    request: {
-      query: CHAT_BY_ID,
-      variables: {
-        id: '1',
-      },
-    },
-    result: {
-      data: {
-        chatById: {
-          _id: '1',
-          isGroup: false,
-          creator: {
-            _id: '64378abd85008f171cf2990d',
-            firstName: 'Wilt',
-            lastName: 'Shepherd',
-            image: null,
-            email: 'testsuperadmin@example.com',
-            createdAt: '2023-04-13T04:53:17.742Z',
-            __typename: 'User',
-          },
-          organization: null,
-          name: '',
-          createdAt: '2345678903456',
-          messages: [
-            {
-              _id: '4',
-              createdAt: '345678908765',
-              messageContent: 'Hello',
-              media: null,
-              replyTo: null,
-              sender: {
-                _id: '2',
-                firstName: 'Test',
-                lastName: 'User',
-                email: 'test@example.com',
-                image: '',
-              },
-            },
-          ],
-          users: [
-            {
-              _id: '1',
-              firstName: 'Disha',
-              lastName: 'Talreja',
-              email: 'disha@example.com',
-              image: '',
-            },
-            {
-              _id: '2',
-              firstName: 'Test',
-              lastName: 'User',
-              email: 'test@example.com',
-              image: '',
-            },
-          ],
-          admins: [],
-          unseenMessagesByUsers: {
-            '1': 0,
-            '2': 0,
-          },
-        },
-      },
-    },
-  },
-  {
-    request: {
-      query: CHAT_BY_ID,
-      variables: {
-        id: '',
-      },
-    },
-    result: {
-      data: {
-        chatById: {
-          _id: '1',
-          isGroup: false,
-          creator: {
-            _id: '64378abd85008f171cf2990d',
-            firstName: 'Wilt',
-            lastName: 'Shepherd',
-            image: null,
-            email: 'testsuperadmin@example.com',
-            createdAt: '2023-04-13T04:53:17.742Z',
-            __typename: 'User',
-          },
-          organization: null,
-          name: '',
-          createdAt: '2345678903456',
-          messages: [
-            {
-              _id: '4',
-              createdAt: '345678908765',
-              messageContent: 'Hello',
-              media: null,
-              replyTo: null,
-              sender: {
-                _id: '2',
-                firstName: 'Test',
-                lastName: 'User',
-                email: 'test@example.com',
-                image: '',
-              },
-            },
-          ],
-          users: [
-            {
-              _id: '1',
-              firstName: 'Disha',
-              lastName: 'Talreja',
-              email: 'disha@example.com',
-              image: '',
-            },
-            {
-              _id: '2',
-              firstName: 'Test',
-              lastName: 'User',
-              email: 'test@example.com',
-              image: '',
-            },
-          ],
-          admins: [],
-          unseenMessagesByUsers: {
-            '1': 0,
-            '2': 0,
-          },
-        },
-      },
-    },
-  },
-];
-
-export const CHATS_LIST_MOCK = [
-  {
-    request: {
-      query: CHATS_LIST,
-      variables: {
-        id: '8',
-      },
-    },
-    result: {
-      data: {
-        chatsByUserId: [
-          {
-            _id: '65844efc814dd40fgh03db811c4',
-            isGroup: true,
-            creator: {
-              _id: '64378abd85008f171cf2990d',
-              firstName: 'Wilt',
-              lastName: 'Shepherd',
-              image: null,
-              email: 'testsuperadmin@example.com',
-              createdAt: '2023-04-13T04:53:17.742Z',
-              __typename: 'User',
-            },
-            organization: {
-              _id: 'pw3ertyuiophgfre45678',
-              name: 'rtyu',
-            },
-            createdAt: '2345678903456',
-            name: 'Test Group Chat',
-            messages: [
-              {
-                _id: '345678',
-                createdAt: '345678908765',
-                messageContent: 'Hello',
-                media: null,
-                replyTo: null,
-                sender: {
-                  _id: '2',
-                  firstName: 'Test',
-                  lastName: 'User',
-                  email: 'test@example.com',
-                  image: '',
-                },
-              },
-            ],
-            users: [
-              {
-                _id: '1',
-                firstName: 'Disha',
-                lastName: 'Talreja',
-                email: 'disha@example.com',
-                image: '',
-              },
-              {
-                _id: '2',
-                firstName: 'Test',
-                lastName: 'User',
-                email: 'test@example.com',
-                image: '',
-              },
-              {
-                _id: '3',
-                firstName: 'Test',
-                lastName: 'User1',
-                email: 'test1@example.com',
-                image: '',
-              },
-              {
-                _id: '4',
-                firstName: 'Test',
-                lastName: 'User2',
-                email: 'test2@example.com',
-                image: '',
-              },
-              {
-                _id: '5',
-                firstName: 'Test',
-                lastName: 'User4',
-                email: 'test4@example.com',
-                image: '',
-              },
-            ],
-            admins: [
-              {
-                _id: '1',
-                firstName: 'Disha',
-                lastName: 'Talreja',
-                email: 'disha@example.com',
-                image: '',
-              },
-            ],
-            unseenMessagesByUsers: {
-              '1': 0,
-              '2': 0,
-              '3': 0,
-              '4': 0,
-              '5': 0,
-            },
-          },
-          {
-            _id: '65844efc814ddgh4003db811c4',
-            isGroup: true,
-            creator: {
-              _id: '64378abd85008f171cf2990d',
-              firstName: 'Wilt',
-              lastName: 'Shepherd',
-              image: null,
-              email: 'testsuperadmin@example.com',
-              createdAt: '2023-04-13T04:53:17.742Z',
-              __typename: 'User',
-            },
-            organization: {
-              _id: 'pw3ertyuiophgfre45678',
-              name: 'rtyu',
-            },
-            createdAt: '2345678903456',
-            name: 'Test Group Chat',
-            messages: [
-              {
-                _id: '345678',
-                createdAt: '345678908765',
-                messageContent: 'Hello',
-                media: null,
-                replyTo: null,
-                sender: {
-                  _id: '2',
-                  firstName: 'Test',
-                  lastName: 'User',
-                  email: 'test@example.com',
-                  image: '',
-                },
-              },
-            ],
-            users: [
-              {
-                _id: '1',
-                firstName: 'Disha',
-                lastName: 'Talreja',
-                email: 'disha@example.com',
-                image: '',
-              },
-              {
-                _id: '2',
-                firstName: 'Test',
-                lastName: 'User',
-                email: 'test@example.com',
-                image: '',
-              },
-              {
-                _id: '3',
-                firstName: 'Test',
-                lastName: 'User1',
-                email: 'test1@example.com',
-                image: '',
-              },
-              {
-                _id: '4',
-                firstName: 'Test',
-                lastName: 'User2',
-                email: 'test2@example.com',
-                image: '',
-              },
-              {
-                _id: '5',
-                firstName: 'Test',
-                lastName: 'User4',
-                email: 'test4@example.com',
-                image: '',
-              },
-            ],
-            unseenMessagesByUsers: {
-              '1': 0,
-              '2': 0,
-              '3': 0,
-              '4': 0,
-              '5': 0,
-            },
-          },
-        ],
-      },
-    },
-  },
-  {
-    request: {
-      query: CHATS_LIST,
-      variables: {
-        id: '8',
-      },
-    },
-    result: {
-      data: {
-        chatsByUserId: [
-          {
-            _id: '65844efc814dd40fgh03db811c4',
-            isGroup: true,
-            creator: {
-              _id: '64378abd85008f171cf2990d',
-              firstName: 'Wilt',
-              lastName: 'Shepherd',
-              image: null,
-              email: 'testsuperadmin@example.com',
-              createdAt: '2023-04-13T04:53:17.742Z',
-              __typename: 'User',
-            },
-            organization: {
-              _id: 'pw3ertyuiophgfre45678',
-              name: 'rtyu',
-            },
-            createdAt: '2345678903456',
-            name: 'Test Group Chat',
-            messages: [
-              {
-                _id: '345678',
-                createdAt: '345678908765',
-                messageContent: 'Hello',
-                media: null,
-                replyTo: null,
-                sender: {
-                  _id: '2',
-                  firstName: 'Test',
-                  lastName: 'User',
-                  email: 'test@example.com',
-                  image: '',
-                },
-              },
-            ],
-            users: [
-              {
-                _id: '1',
-                firstName: 'Disha',
-                lastName: 'Talreja',
-                email: 'disha@example.com',
-                image: '',
-              },
-              {
-                _id: '2',
-                firstName: 'Test',
-                lastName: 'User',
-                email: 'test@example.com',
-                image: '',
-              },
-              {
-                _id: '3',
-                firstName: 'Test',
-                lastName: 'User1',
-                email: 'test1@example.com',
-                image: '',
-              },
-              {
-                _id: '4',
-                firstName: 'Test',
-                lastName: 'User2',
-                email: 'test2@example.com',
-                image: '',
-              },
-              {
-                _id: '5',
-                firstName: 'Test',
-                lastName: 'User4',
-                email: 'test4@example.com',
-                image: '',
-              },
-            ],
-            admins: [
-              {
-                _id: '1',
-                firstName: 'Disha',
-                lastName: 'Talreja',
-                email: 'disha@example.com',
-                image: '',
-              },
-            ],
-            unseenMessagesByUsers: {
-              '1': 0,
-              '2': 0,
-              '3': 0,
-              '4': 0,
-              '5': 0,
-            },
-          },
-          {
-            _id: '65844efc814ddgh4003db811c4',
-            isGroup: true,
-            creator: {
-              _id: '64378abd85008f171cf2990d',
-              firstName: 'Wilt',
-              lastName: 'Shepherd',
-              image: null,
-              email: 'testsuperadmin@example.com',
-              createdAt: '2023-04-13T04:53:17.742Z',
-              __typename: 'User',
-            },
-            organization: {
-              _id: 'pw3ertyuiophgfre45678',
-              name: 'rtyu',
-            },
-            createdAt: '2345678903456',
-            name: 'Test Group Chat',
-            messages: [
-              {
-                _id: '345678',
-                createdAt: '345678908765',
-                messageContent: 'Hello',
-                media: null,
-                replyTo: null,
-                sender: {
-                  _id: '2',
-                  firstName: 'Test',
-                  lastName: 'User',
-                  email: 'test@example.com',
-                  image: '',
-                },
-              },
-            ],
-            users: [
-              {
-                _id: '1',
-                firstName: 'Disha',
-                lastName: 'Talreja',
-                email: 'disha@example.com',
-                image: '',
-              },
-              {
-                _id: '2',
-                firstName: 'Test',
-                lastName: 'User',
-                email: 'test@example.com',
-                image: '',
-              },
-              {
-                _id: '3',
-                firstName: 'Test',
-                lastName: 'User1',
-                email: 'test1@example.com',
-                image: '',
-              },
-              {
-                _id: '4',
-                firstName: 'Test',
-                lastName: 'User2',
-                email: 'test2@example.com',
-                image: '',
-              },
-              {
-                _id: '5',
-                firstName: 'Test',
-                lastName: 'User4',
-                email: 'test4@example.com',
-                image: '',
-              },
-            ],
-            admins: [
-              {
-                _id: '1',
-                firstName: 'Disha',
-                lastName: 'Talreja',
-                email: 'disha@example.com',
-                image: '',
-              },
-            ],
-            unseenMessagesByUsers: {
-              '1': 0,
-              '2': 0,
-              '3': 0,
-              '4': 0,
-              '5': 0,
-            },
-          },
-        ],
-      },
-    },
-  },
-  {
-    request: {
-      query: CHATS_LIST,
-      variables: {
-        id: '8',
-      },
-    },
-    result: {
-      data: {
-        chatsByUserId: [
-          {
-            _id: '65844efc814dd40fgh03db811c4',
-            isGroup: true,
-            creator: {
-              _id: '64378abd85008f171cf2990d',
-              firstName: 'Wilt',
-              lastName: 'Shepherd',
-              image: null,
-              email: 'testsuperadmin@example.com',
-              createdAt: '2023-04-13T04:53:17.742Z',
-              __typename: 'User',
-            },
-            organization: {
-              _id: 'pw3ertyuiophgfre45678',
-              name: 'rtyu',
-            },
-            createdAt: '2345678903456',
-            name: 'Test Group Chat',
-            messages: [
-              {
-                _id: '345678',
-                createdAt: '345678908765',
-                messageContent: 'Hello',
-                media: null,
-                replyTo: null,
-                sender: {
-                  _id: '2',
-                  firstName: 'Test',
-                  lastName: 'User',
-                  email: 'test@example.com',
-                  image: '',
-                },
-              },
-            ],
-            users: [
-              {
-                _id: '1',
-                firstName: 'Disha',
-                lastName: 'Talreja',
-                email: 'disha@example.com',
-                image: '',
-              },
-              {
-                _id: '2',
-                firstName: 'Test',
-                lastName: 'User',
-                email: 'test@example.com',
-                image: '',
-              },
-              {
-                _id: '3',
-                firstName: 'Test',
-                lastName: 'User1',
-                email: 'test1@example.com',
-                image: '',
-              },
-              {
-                _id: '4',
-                firstName: 'Test',
-                lastName: 'User2',
-                email: 'test2@example.com',
-                image: '',
-              },
-              {
-                _id: '5',
-                firstName: 'Test',
-                lastName: 'User4',
-                email: 'test4@example.com',
-                image: '',
-              },
-            ],
-            admins: [
-              {
-                _id: '1',
-                firstName: 'Disha',
-                lastName: 'Talreja',
-                email: 'disha@example.com',
-                image: '',
-              },
-            ],
-            unseenMessagesByUsers: {
-              '1': 0,
-              '2': 0,
-              '3': 0,
-              '4': 0,
-              '5': 0,
-            },
-          },
-          {
-            _id: '65844efc814ddgh4003db811c4',
-            isGroup: true,
-            creator: {
-              _id: '64378abd85008f171cf2990d',
-              firstName: 'Wilt',
-              lastName: 'Shepherd',
-              image: null,
-              email: 'testsuperadmin@example.com',
-              createdAt: '2023-04-13T04:53:17.742Z',
-              __typename: 'User',
-            },
-            organization: {
-              _id: 'pw3ertyuiophgfre45678',
-              name: 'rtyu',
-            },
-            createdAt: '2345678903456',
-            name: 'Test Group Chat',
-            messages: [
-              {
-                _id: '345678',
-                createdAt: '345678908765',
-                messageContent: 'Hello',
-                media: null,
-                replyTo: null,
-                sender: {
-                  _id: '2',
-                  firstName: 'Test',
-                  lastName: 'User',
-                  email: 'test@example.com',
-                  image: '',
-                },
-              },
-            ],
-            users: [
-              {
-                _id: '1',
-                firstName: 'Disha',
-                lastName: 'Talreja',
-                email: 'disha@example.com',
-                image: '',
-              },
-              {
-                _id: '2',
-                firstName: 'Test',
-                lastName: 'User',
-                email: 'test@example.com',
-                image: '',
-              },
-              {
-                _id: '3',
-                firstName: 'Test',
-                lastName: 'User1',
-                email: 'test1@example.com',
-                image: '',
-              },
-              {
-                _id: '4',
-                firstName: 'Test',
-                lastName: 'User2',
-                email: 'test2@example.com',
-                image: '',
-              },
-              {
-                _id: '5',
-                firstName: 'Test',
-                lastName: 'User4',
-                email: 'test4@example.com',
-                image: '',
-              },
-            ],
-            admins: [
-              {
-                _id: '1',
-                firstName: 'Disha',
-                lastName: 'Talreja',
-                email: 'disha@example.com',
-                image: '',
-              },
-            ],
-            unseenMessagesByUsers: {
-              '1': 0,
-              '2': 0,
-              '3': 0,
-              '4': 0,
-              '5': 0,
-            },
-          },
-        ],
-      },
-    },
-  },
-  {
-    request: {
-      query: CHATS_LIST,
-      variables: {
-        id: null,
-      },
-    },
-    result: {
-      data: {
-        chatsByUserId: [
-          {
-            _id: '65844efc814dd40fgh03db811c4',
-            isGroup: true,
-            creator: {
-              _id: '64378abd85008f171cf2990d',
-              firstName: 'Wilt',
-              lastName: 'Shepherd',
-              image: null,
-              email: 'testsuperadmin@example.com',
-              createdAt: '2023-04-13T04:53:17.742Z',
-              __typename: 'User',
-            },
-            organization: {
-              _id: 'pw3ertyuiophgfre45678',
-              name: 'rtyu',
-            },
-            createdAt: '2345678903456',
-            name: 'Test Group Chat',
-            messages: [
-              {
-                _id: '345678',
-                createdAt: '345678908765',
-                messageContent: 'Hello',
-                media: null,
-                replyTo: null,
-                sender: {
-                  _id: '2',
-                  firstName: 'Test',
-                  lastName: 'User',
-                  email: 'test@example.com',
-                  image: '',
-                },
-              },
-            ],
-            users: [
-              {
-                _id: '1',
-                firstName: 'Disha',
-                lastName: 'Talreja',
-                email: 'disha@example.com',
-                image: '',
-              },
-              {
-                _id: '2',
-                firstName: 'Test',
-                lastName: 'User',
-                email: 'test@example.com',
-                image: '',
-              },
-              {
-                _id: '3',
-                firstName: 'Test',
-                lastName: 'User1',
-                email: 'test1@example.com',
-                image: '',
-              },
-              {
-                _id: '4',
-                firstName: 'Test',
-                lastName: 'User2',
-                email: 'test2@example.com',
-                image: '',
-              },
-              {
-                _id: '5',
-                firstName: 'Test',
-                lastName: 'User4',
-                email: 'test4@example.com',
-                image: '',
-              },
-            ],
-            admins: [
-              {
-                _id: '1',
-                firstName: 'Disha',
-                lastName: 'Talreja',
-                email: 'disha@example.com',
-                image: '',
-              },
-            ],
-            unseenMessagesByUsers: {
-              '1': 0,
-              '2': 0,
-              '3': 0,
-              '4': 0,
-              '5': 0,
-            },
-          },
-          {
-            _id: '65844efc814ddgh4003db811c4',
-            isGroup: true,
-            creator: {
-              _id: '64378abd85008f171cf2990d',
-              firstName: 'Wilt',
-              lastName: 'Shepherd',
-              image: null,
-              email: 'testsuperadmin@example.com',
-              createdAt: '2023-04-13T04:53:17.742Z',
-              __typename: 'User',
-            },
-            organization: {
-              _id: 'pw3ertyuiophgfre45678',
-              name: 'rtyu',
-            },
-            createdAt: '2345678903456',
-            name: 'Test Group Chat',
-            messages: [
-              {
-                _id: '345678',
-                createdAt: '345678908765',
-                messageContent: 'Hello',
-                media: null,
-                replyTo: null,
-                sender: {
-                  _id: '2',
-                  firstName: 'Test',
-                  lastName: 'User',
-                  email: 'test@example.com',
-                  image: '',
-                },
-              },
-            ],
-            users: [
-              {
-                _id: '1',
-                firstName: 'Disha',
-                lastName: 'Talreja',
-                email: 'disha@example.com',
-                image: '',
-              },
-              {
-                _id: '2',
-                firstName: 'Test',
-                lastName: 'User',
-                email: 'test@example.com',
-                image: '',
-              },
-              {
-                _id: '3',
-                firstName: 'Test',
-                lastName: 'User1',
-                email: 'test1@example.com',
-                image: '',
-              },
-              {
-                _id: '4',
-                firstName: 'Test',
-                lastName: 'User2',
-                email: 'test2@example.com',
-                image: '',
-              },
-              {
-                _id: '5',
-                firstName: 'Test',
-                lastName: 'User4',
-                email: 'test4@example.com',
-                image: '',
-              },
-            ],
-            admins: [
-              {
-                _id: '1',
-                firstName: 'Disha',
-                lastName: 'Talreja',
-                email: 'disha@example.com',
-                image: '',
-              },
-            ],
-            unseenMessagesByUsers: {
-              '1': 0,
-              '2': 0,
-              '3': 0,
-              '4': 0,
-              '5': 0,
-            },
-          },
-        ],
-      },
-    },
-  },
-  {
-    request: {
-      query: CHATS_LIST,
-      variables: {
-        id: null,
-      },
-    },
-    result: {
-      data: {
-        chatsByUserId: [
-          {
-            _id: '65844efc814dd40fgh03db811c4',
-            isGroup: true,
-            creator: {
-              _id: '64378abd85008f171cf2990d',
-              firstName: 'Wilt',
-              lastName: 'Shepherd',
-              image: null,
-              email: 'testsuperadmin@example.com',
-              createdAt: '2023-04-13T04:53:17.742Z',
-              __typename: 'User',
-            },
-            organization: {
-              _id: 'pw3ertyuiophgfre45678',
-              name: 'rtyu',
-            },
-            createdAt: '2345678903456',
-            name: 'Test Group Chat',
-            messages: [
-              {
-                _id: '345678',
-                createdAt: '345678908765',
-                messageContent: 'Hello',
-                media: null,
-                replyTo: null,
-                type: 'STRING',
-                sender: {
-                  _id: '2',
-                  firstName: 'Test',
-                  lastName: 'User',
-                  email: 'test@example.com',
-                  image: '',
-                },
-              },
-            ],
-            users: [
-              {
-                _id: '1',
-                firstName: 'Disha',
-                lastName: 'Talreja',
-                email: 'disha@example.com',
-                image: '',
-              },
-              {
-                _id: '2',
-                firstName: 'Test',
-                lastName: 'User',
-                email: 'test@example.com',
-                image: '',
-              },
-              {
-                _id: '3',
-                firstName: 'Test',
-                lastName: 'User1',
-                email: 'test1@example.com',
-                image: '',
-              },
-              {
-                _id: '4',
-                firstName: 'Test',
-                lastName: 'User2',
-                email: 'test2@example.com',
-                image: '',
-              },
-              {
-                _id: '5',
-                firstName: 'Test',
-                lastName: 'User4',
-                email: 'test4@example.com',
-                image: '',
-              },
-            ],
-            admins: [
-              {
-                _id: '1',
-                firstName: 'Disha',
-                lastName: 'Talreja',
-                email: 'disha@example.com',
-                image: '',
-              },
-            ],
-            unseenMessagesByUsers: {
-              '1': 0,
-              '2': 0,
-              '3': 0,
-              '4': 0,
-              '5': 0,
-            },
-          },
-          {
-            _id: '65844efc814ddgh4003db811c4',
-            isGroup: true,
-            creator: {
-              _id: '64378abd85008f171cf2990d',
-              firstName: 'Wilt',
-              lastName: 'Shepherd',
-              image: null,
-              email: 'testsuperadmin@example.com',
-              createdAt: '2023-04-13T04:53:17.742Z',
-              __typename: 'User',
-            },
-            organization: {
-              _id: 'pw3ertyuiophgfre45678',
-              name: 'rtyu',
-            },
-            createdAt: '2345678903456',
-            name: 'Test Group Chat',
-            messages: [
-              {
-                _id: '345678',
-                createdAt: '345678908765',
-                messageContent: 'Hello',
-                media: null,
-                replyTo: null,
-                type: 'STRING',
-                sender: {
-                  _id: '2',
-                  firstName: 'Test',
-                  lastName: 'User',
-                  email: 'test@example.com',
-                  image: '',
-                },
-              },
-            ],
-            users: [
-              {
-                _id: '1',
-                firstName: 'Disha',
-                lastName: 'Talreja',
-                email: 'disha@example.com',
-                image: '',
-              },
-              {
-                _id: '2',
-                firstName: 'Test',
-                lastName: 'User',
-                email: 'test@example.com',
-                image: '',
-              },
-              {
-                _id: '3',
-                firstName: 'Test',
-                lastName: 'User1',
-                email: 'test1@example.com',
-                image: '',
-              },
-              {
-                _id: '4',
-                firstName: 'Test',
-                lastName: 'User2',
-                email: 'test2@example.com',
-                image: '',
-              },
-              {
-                _id: '5',
-                firstName: 'Test',
-                lastName: 'User4',
-                email: 'test4@example.com',
-                image: '',
-              },
-            ],
-            admins: [
-              {
-                _id: '1',
-                firstName: 'Disha',
-                lastName: 'Talreja',
-                email: 'disha@example.com',
-                image: '',
-              },
-            ],
-            unseenMessagesByUsers: {
-              '1': 0,
-              '2': 0,
-              '3': 0,
-              '4': 0,
-              '5': 0,
-            },
-          },
-        ],
-      },
-    },
-  },
-  {
-    request: {
-      query: CHATS_LIST,
-      variables: {
-        id: null,
-      },
-    },
-    result: {
-      data: {
-        chatsByUserId: [
-          {
-            _id: '65844efc814dd40fgh03db811c4',
-            isGroup: true,
-            creator: {
-              _id: '64378abd85008f171cf2990d',
-              firstName: 'Wilt',
-              lastName: 'Shepherd',
-              image: null,
-              email: 'testsuperadmin@example.com',
-              createdAt: '2023-04-13T04:53:17.742Z',
-              __typename: 'User',
-            },
-            organization: {
-              _id: 'pw3ertyuiophgfre45678',
-              name: 'rtyu',
-            },
-            createdAt: '2345678903456',
-            name: 'Test Group Chat',
-            messages: [
-              {
-                _id: '345678',
-                createdAt: '345678908765',
-                messageContent: 'Hello',
-                media: null,
-                replyTo: null,
-                type: 'STRING',
-                sender: {
-                  _id: '2',
-                  firstName: 'Test',
-                  lastName: 'User',
-                  email: 'test@example.com',
-                  image: '',
-                },
-              },
-            ],
-            users: [
-              {
-                _id: '1',
-                firstName: 'Disha',
-                lastName: 'Talreja',
-                email: 'disha@example.com',
-                image: '',
-              },
-              {
-                _id: '2',
-                firstName: 'Test',
-                lastName: 'User',
-                email: 'test@example.com',
-                image: '',
-              },
-              {
-                _id: '3',
-                firstName: 'Test',
-                lastName: 'User1',
-                email: 'test1@example.com',
-                image: '',
-              },
-              {
-                _id: '4',
-                firstName: 'Test',
-                lastName: 'User2',
-                email: 'test2@example.com',
-                image: '',
-              },
-              {
-                _id: '5',
-                firstName: 'Test',
-                lastName: 'User4',
-                email: 'test4@example.com',
-                image: '',
-              },
-            ],
-            admins: [
-              {
-                _id: '1',
-                firstName: 'Disha',
-                lastName: 'Talreja',
-                email: 'disha@example.com',
-                image: '',
-              },
-            ],
-            unseenMessagesByUsers: {
-              '1': 0,
-              '2': 0,
-              '3': 0,
-              '4': 0,
-              '5': 0,
-            },
-          },
-          {
-            _id: '65844efc814ddgh4003db811c4',
-            isGroup: true,
-            creator: {
-              _id: '64378abd85008f171cf2990d',
-              firstName: 'Wilt',
-              lastName: 'Shepherd',
-              image: null,
-              email: 'testsuperadmin@example.com',
-              createdAt: '2023-04-13T04:53:17.742Z',
-              __typename: 'User',
-            },
-            organization: {
-              _id: 'pw3ertyuiophgfre45678',
-              name: 'rtyu',
-            },
-            createdAt: '2345678903456',
-            name: 'Test Group Chat',
-            messages: [
-              {
-                _id: '345678',
-                createdAt: '345678908765',
-                messageContent: 'Hello',
-                media: null,
-                replyTo: null,
-                type: 'STRING',
-                sender: {
-                  _id: '2',
-                  firstName: 'Test',
-                  lastName: 'User',
-                  email: 'test@example.com',
-                  image: '',
-                },
-              },
-            ],
-            users: [
-              {
-                _id: '1',
-                firstName: 'Disha',
-                lastName: 'Talreja',
-                email: 'disha@example.com',
-                image: '',
-              },
-              {
-                _id: '2',
-                firstName: 'Test',
-                lastName: 'User',
-                email: 'test@example.com',
-                image: '',
-              },
-              {
-                _id: '3',
-                firstName: 'Test',
-                lastName: 'User1',
-                email: 'test1@example.com',
-                image: '',
-              },
-              {
-                _id: '4',
-                firstName: 'Test',
-                lastName: 'User2',
-                email: 'test2@example.com',
-                image: '',
-              },
-              {
-                _id: '5',
-                firstName: 'Test',
-                lastName: 'User4',
-                email: 'test4@example.com',
-                image: '',
-              },
-            ],
-            admins: [
-              {
-                _id: '1',
-                firstName: 'Disha',
-                lastName: 'Talreja',
-                email: 'disha@example.com',
-                image: '',
-              },
-            ],
-            unseenMessagesByUsers: {
-              '1': 0,
-              '2': 0,
-              '3': 0,
-              '4': 0,
-              '5': 0,
-            },
-          },
-        ],
-      },
-    },
-  },
-  {
-    request: {
-      query: CHATS_LIST,
-      variables: {
-        id: null,
-      },
-    },
-    result: {
-      data: {
-        chatsByUserId: [
-          {
-            _id: '65844efc814dd40fgh03db811c4',
-            isGroup: true,
-            creator: {
-              _id: '64378abd85008f171cf2990d',
-              firstName: 'Wilt',
-              lastName: 'Shepherd',
-              image: null,
-              email: 'testsuperadmin@example.com',
-              createdAt: '2023-04-13T04:53:17.742Z',
-              __typename: 'User',
-            },
-            organization: {
-              _id: 'pw3ertyuiophgfre45678',
-              name: 'rtyu',
-            },
-            createdAt: '2345678903456',
-            name: 'Test Group Chat',
-            messages: [
-              {
-                _id: '345678',
-                createdAt: '345678908765',
-                messageContent: 'Hello',
-                media: null,
-                replyTo: null,
-                type: 'STRING',
-                sender: {
-                  _id: '2',
-                  firstName: 'Test',
-                  lastName: 'User',
-                  email: 'test@example.com',
-                  image: '',
-                },
-              },
-            ],
-            users: [
-              {
-                _id: '1',
-                firstName: 'Disha',
-                lastName: 'Talreja',
-                email: 'disha@example.com',
-                image: '',
-              },
-              {
-                _id: '2',
-                firstName: 'Test',
-                lastName: 'User',
-                email: 'test@example.com',
-                image: '',
-              },
-              {
-                _id: '3',
-                firstName: 'Test',
-                lastName: 'User1',
-                email: 'test1@example.com',
-                image: '',
-              },
-              {
-                _id: '4',
-                firstName: 'Test',
-                lastName: 'User2',
-                email: 'test2@example.com',
-                image: '',
-              },
-              {
-                _id: '5',
-                firstName: 'Test',
-                lastName: 'User4',
-                email: 'test4@example.com',
-                image: '',
-              },
-            ],
-            admins: [
-              {
-                _id: '1',
-                firstName: 'Disha',
-                lastName: 'Talreja',
-                email: 'disha@example.com',
-                image: '',
-              },
-            ],
-            unseenMessagesByUsers: {
-              '1': 0,
-              '2': 0,
-              '3': 0,
-              '4': 0,
-              '5': 0,
-            },
-          },
-          {
-            _id: '65844efc814ddgh4003db811c4',
-            isGroup: true,
-            creator: {
-              _id: '64378abd85008f171cf2990d',
-              firstName: 'Wilt',
-              lastName: 'Shepherd',
-              image: null,
-              email: 'testsuperadmin@example.com',
-              createdAt: '2023-04-13T04:53:17.742Z',
-              __typename: 'User',
-            },
-            organization: {
-              _id: 'pw3ertyuiophgfre45678',
-              name: 'rtyu',
-            },
-            createdAt: '2345678903456',
-            name: 'Test Group Chat',
-            messages: [
-              {
-                _id: '345678',
-                createdAt: '345678908765',
-                messageContent: 'Hello',
-                media: null,
-                replyTo: null,
-                type: 'STRING',
-                sender: {
-                  _id: '2',
-                  firstName: 'Test',
-                  lastName: 'User',
-                  email: 'test@example.com',
-                  image: '',
-                },
-              },
-            ],
-            users: [
-              {
-                _id: '1',
-                firstName: 'Disha',
-                lastName: 'Talreja',
-                email: 'disha@example.com',
-                image: '',
-              },
-              {
-                _id: '2',
-                firstName: 'Test',
-                lastName: 'User',
-                email: 'test@example.com',
-                image: '',
-              },
-              {
-                _id: '3',
-                firstName: 'Test',
-                lastName: 'User1',
-                email: 'test1@example.com',
-                image: '',
-              },
-              {
-                _id: '4',
-                firstName: 'Test',
-                lastName: 'User2',
-                email: 'test2@example.com',
-                image: '',
-              },
-              {
-                _id: '5',
-                firstName: 'Test',
-                lastName: 'User4',
-                email: 'test4@example.com',
-                image: '',
-              },
-            ],
-            admins: [
-              {
-                _id: '1',
-                firstName: 'Disha',
-                lastName: 'Talreja',
-                email: 'disha@example.com',
-                image: '',
-              },
-            ],
-            unseenMessagesByUsers: {
-              '1': 0,
-              '2': 0,
-              '3': 0,
-              '4': 0,
-              '5': 0,
-            },
-          },
-        ],
-      },
-    },
-  },
-  {
-    request: {
-      query: CHATS_LIST,
-      variables: {
-        id: null,
-      },
-    },
-    result: {
-      data: {
-        chatsByUserId: [
-          {
-            _id: '65844efc814dd40fgh03db811c4',
-            isGroup: true,
-            creator: {
-              _id: '64378abd85008f171cf2990d',
-              firstName: 'Wilt',
-              lastName: 'Shepherd',
-              image: null,
-              email: 'testsuperadmin@example.com',
-              createdAt: '2023-04-13T04:53:17.742Z',
-              __typename: 'User',
-            },
-            organization: {
-              _id: 'pw3ertyuiophgfre45678',
-              name: 'rtyu',
-            },
-            createdAt: '2345678903456',
-            name: 'Test Group Chat',
-            messages: [
-              {
-                _id: '345678',
-                createdAt: '345678908765',
-                messageContent: 'Hello',
-                media: null,
-                replyTo: null,
-                type: 'STRING',
-                sender: {
-                  _id: '2',
-                  firstName: 'Test',
-                  lastName: 'User',
-                  email: 'test@example.com',
-                  image: '',
-                },
-              },
-            ],
-            users: [
-              {
-                _id: '1',
-                firstName: 'Disha',
-                lastName: 'Talreja',
-                email: 'disha@example.com',
-                image: '',
-              },
-              {
-                _id: '2',
-                firstName: 'Test',
-                lastName: 'User',
-                email: 'test@example.com',
-                image: '',
-              },
-              {
-                _id: '3',
-                firstName: 'Test',
-                lastName: 'User1',
-                email: 'test1@example.com',
-                image: '',
-              },
-              {
-                _id: '4',
-                firstName: 'Test',
-                lastName: 'User2',
-                email: 'test2@example.com',
-                image: '',
-              },
-              {
-                _id: '5',
-                firstName: 'Test',
-                lastName: 'User4',
-                email: 'test4@example.com',
-                image: '',
-              },
-            ],
-            admins: [
-              {
-                _id: '1',
-                firstName: 'Disha',
-                lastName: 'Talreja',
-                email: 'disha@example.com',
-                image: '',
-              },
-            ],
-            unseenMessagesByUsers: {
-              '1': 0,
-              '2': 0,
-              '3': 0,
-              '4': 0,
-              '5': 0,
-            },
-          },
-          {
-            _id: '65844efc814ddgh4003db811c4',
-            isGroup: true,
-            creator: {
-              _id: '64378abd85008f171cf2990d',
-              firstName: 'Wilt',
-              lastName: 'Shepherd',
-              image: null,
-              email: 'testsuperadmin@example.com',
-              createdAt: '2023-04-13T04:53:17.742Z',
-              __typename: 'User',
-            },
-            organization: {
-              _id: 'pw3ertyuiophgfre45678',
-              name: 'rtyu',
-            },
-            createdAt: '2345678903456',
-            name: 'Test Group Chat',
-            messages: [
-              {
-                _id: '345678',
-                createdAt: '345678908765',
-                messageContent: 'Hello',
-                media: null,
-                replyTo: null,
-                type: 'STRING',
-                sender: {
-                  _id: '2',
-                  firstName: 'Test',
-                  lastName: 'User',
-                  email: 'test@example.com',
-                  image: '',
-                },
-              },
-            ],
-            users: [
-              {
-                _id: '1',
-                firstName: 'Disha',
-                lastName: 'Talreja',
-                email: 'disha@example.com',
-                image: '',
-              },
-              {
-                _id: '2',
-                firstName: 'Test',
-                lastName: 'User',
-                email: 'test@example.com',
-                image: '',
-              },
-              {
-                _id: '3',
-                firstName: 'Test',
-                lastName: 'User1',
-                email: 'test1@example.com',
-                image: '',
-              },
-              {
-                _id: '4',
-                firstName: 'Test',
-                lastName: 'User2',
-                email: 'test2@example.com',
-                image: '',
-              },
-              {
-                _id: '5',
-                firstName: 'Test',
-                lastName: 'User4',
-                email: 'test4@example.com',
-                image: '',
-              },
-            ],
-            admins: [
-              {
-                _id: '1',
-                firstName: 'Disha',
-                lastName: 'Talreja',
-                email: 'disha@example.com',
-                image: '',
-              },
-            ],
-            unseenMessagesByUsers: {
-              '1': 0,
-              '2': 0,
-              '3': 0,
-              '4': 0,
-              '5': 0,
-            },
-          },
-        ],
-      },
-    },
-  },
-  {
-    request: {
-      query: CHATS_LIST,
-      variables: {
-        id: '2',
-      },
-    },
-    result: {
-      data: {
-        chatsByUserId: [
-          {
-            _id: '65844efc814dd40fgh03db811c4',
-            isGroup: true,
-            creator: {
-              _id: '64378abd85008f171cf2990d',
-              firstName: 'Wilt',
-              lastName: 'Shepherd',
-              image: null,
-              email: 'testsuperadmin@example.com',
-              createdAt: '2023-04-13T04:53:17.742Z',
-              __typename: 'User',
-            },
-            organization: {
-              _id: 'pw3ertyuiophgfre45678',
-              name: 'rtyu',
-            },
-            createdAt: '2345678903456',
-            name: 'Test Group Chat',
-            messages: [
-              {
-                _id: '345678',
-                createdAt: '345678908765',
-                messageContent: 'Hello',
-                media: null,
-                replyTo: null,
-                type: 'STRING',
-                sender: {
-                  _id: '2',
-                  firstName: 'Test',
-                  lastName: 'User',
-                  email: 'test@example.com',
-                  image: '',
-                },
-              },
-            ],
-            users: [
-              {
-                _id: '1',
-                firstName: 'Disha',
-                lastName: 'Talreja',
-                email: 'disha@example.com',
-                image: '',
-              },
-              {
-                _id: '2',
-                firstName: 'Test',
-                lastName: 'User',
-                email: 'test@example.com',
-                image: '',
-              },
-              {
-                _id: '3',
-                firstName: 'Test',
-                lastName: 'User1',
-                email: 'test1@example.com',
-                image: '',
-              },
-              {
-                _id: '4',
-                firstName: 'Test',
-                lastName: 'User2',
-                email: 'test2@example.com',
-                image: '',
-              },
-              {
-                _id: '5',
-                firstName: 'Test',
-                lastName: 'User4',
-                email: 'test4@example.com',
-                image: '',
-              },
-            ],
-            admins: [
-              {
-                _id: '1',
-                firstName: 'Disha',
-                lastName: 'Talreja',
-                email: 'disha@example.com',
-                image: '',
-              },
-            ],
-            unseenMessagesByUsers: JSON.stringify({
-              '1': 0,
-              '2': 0,
-              '3': 0,
-              '4': 0,
-              '5': 0,
-            }),
-          },
-          {
-            _id: '65844efc814ddgh4003db811c4',
-            isGroup: true,
-            creator: {
-              _id: '64378abd85008f171cf2990d',
-              firstName: 'Wilt',
-              lastName: 'Shepherd',
-              image: null,
-              email: 'testsuperadmin@example.com',
-              createdAt: '2023-04-13T04:53:17.742Z',
-              __typename: 'User',
-            },
-            organization: {
-              _id: 'pw3ertyuiophgfre45678',
-              name: 'rtyu',
-            },
-            createdAt: '2345678903456',
-            name: 'Test Group Chat',
-            messages: [
-              {
-                _id: '345678',
-                createdAt: '345678908765',
-                messageContent: 'Hello',
-                media: null,
-                replyTo: null,
-                type: 'STRING',
-                sender: {
-                  _id: '2',
-                  firstName: 'Test',
-                  lastName: 'User',
-                  email: 'test@example.com',
-                  image: '',
-                },
-              },
-            ],
-            users: [
-              {
-                _id: '1',
-                firstName: 'Disha',
-                lastName: 'Talreja',
-                email: 'disha@example.com',
-                image: '',
-              },
-              {
-                _id: '2',
-                firstName: 'Test',
-                lastName: 'User',
-                email: 'test@example.com',
-                image: '',
-              },
-              {
-                _id: '3',
-                firstName: 'Test',
-                lastName: 'User1',
-                email: 'test1@example.com',
-                image: '',
-              },
-              {
-                _id: '4',
-                firstName: 'Test',
-                lastName: 'User2',
-                email: 'test2@example.com',
-                image: '',
-              },
-              {
-                _id: '5',
-                firstName: 'Test',
-                lastName: 'User4',
-                email: 'test4@example.com',
-                image: '',
-              },
-            ],
-            admins: [
-              {
-                _id: '1',
-                firstName: 'Disha',
-                lastName: 'Talreja',
-                email: 'disha@example.com',
-                image: '',
-              },
-            ],
-            unseenMessagesByUsers: JSON.stringify({
-              '1': 0,
-              '2': 0,
-              '3': 0,
-              '4': 0,
-              '5': 0,
-            }),
-          },
-        ],
-      },
-    },
-  },
-  {
-    request: {
-      query: CHATS_LIST,
-      variables: {
-        id: '2',
-      },
-    },
-    result: {
-      data: {
-        chatsByUserId: [
-          {
-            _id: '65844efc814dd40fgh03db811c4',
-            isGroup: true,
-            creator: {
-              _id: '64378abd85008f171cf2990d',
-              firstName: 'Wilt',
-              lastName: 'Shepherd',
-              image: null,
-              email: 'testsuperadmin@example.com',
-              createdAt: '2023-04-13T04:53:17.742Z',
-              __typename: 'User',
-            },
-            organization: {
-              _id: 'pw3ertyuiophgfre45678',
-              name: 'rtyu',
-            },
-            createdAt: '2345678903456',
-            name: 'Test Group Chat',
-            messages: [
-              {
-                _id: '345678',
-                createdAt: '345678908765',
-                messageContent: 'Hello',
-                media: null,
-                replyTo: null,
-                type: 'STRING',
-                sender: {
-                  _id: '2',
-                  firstName: 'Test',
-                  lastName: 'User',
-                  email: 'test@example.com',
-                  image: '',
-                },
-              },
-            ],
-            users: [
-              {
-                _id: '1',
-                firstName: 'Disha',
-                lastName: 'Talreja',
-                email: 'disha@example.com',
-                image: '',
-              },
-              {
-                _id: '2',
-                firstName: 'Test',
-                lastName: 'User',
-                email: 'test@example.com',
-                image: '',
-              },
-              {
-                _id: '3',
-                firstName: 'Test',
-                lastName: 'User1',
-                email: 'test1@example.com',
-                image: '',
-              },
-              {
-                _id: '4',
-                firstName: 'Test',
-                lastName: 'User2',
-                email: 'test2@example.com',
-                image: '',
-              },
-              {
-                _id: '5',
-                firstName: 'Test',
-                lastName: 'User4',
-                email: 'test4@example.com',
-                image: '',
-              },
-            ],
-            admins: [
-              {
-                _id: '1',
-                firstName: 'Disha',
-                lastName: 'Talreja',
-                email: 'disha@example.com',
-                image: '',
-              },
-            ],
-            unseenMessagesByUsers: {
-              '1': 0,
-              '2': 0,
-              '3': 0,
-              '4': 0,
-              '5': 0,
-            },
-          },
-          {
-            _id: '65844efc814ddgh4003db811c4',
-            isGroup: true,
-            creator: {
-              _id: '64378abd85008f171cf2990d',
-              firstName: 'Wilt',
-              lastName: 'Shepherd',
-              image: null,
-              email: 'testsuperadmin@example.com',
-              createdAt: '2023-04-13T04:53:17.742Z',
-              __typename: 'User',
-            },
-            organization: {
-              _id: 'pw3ertyuiophgfre45678',
-              name: 'rtyu',
-            },
-            createdAt: '2345678903456',
-            name: 'Test Group Chat',
-            messages: [
-              {
-                _id: '345678',
-                createdAt: '345678908765',
-                messageContent: 'Hello',
-                media: null,
-                replyTo: null,
-                type: 'STRING',
-                sender: {
-                  _id: '2',
-                  firstName: 'Test',
-                  lastName: 'User',
-                  email: 'test@example.com',
-                  image: '',
-                },
-              },
-            ],
-            users: [
-              {
-                _id: '1',
-                firstName: 'Disha',
-                lastName: 'Talreja',
-                email: 'disha@example.com',
-                image: '',
-              },
-              {
-                _id: '2',
-                firstName: 'Test',
-                lastName: 'User',
-                email: 'test@example.com',
-                image: '',
-              },
-              {
-                _id: '3',
-                firstName: 'Test',
-                lastName: 'User1',
-                email: 'test1@example.com',
-                image: '',
-              },
-              {
-                _id: '4',
-                firstName: 'Test',
-                lastName: 'User2',
-                email: 'test2@example.com',
-                image: '',
-              },
-              {
-                _id: '5',
-                firstName: 'Test',
-                lastName: 'User4',
-                email: 'test4@example.com',
-                image: '',
-              },
-            ],
-            admins: [
-              {
-                _id: '1',
-                firstName: 'Disha',
-                lastName: 'Talreja',
-                email: 'disha@example.com',
-                image: '',
-              },
-            ],
-            unseenMessagesByUsers: {
-              '1': 0,
-              '2': 0,
-              '3': 0,
-              '4': 0,
-              '5': 0,
-            },
-          },
-        ],
-      },
-    },
-  },
-  {
-    request: {
-      query: CHATS_LIST,
-      variables: {
-        id: '2',
-      },
-    },
-    result: {
-      data: {
-        chatsByUserId: [
-          {
-            _id: '65844efc814dd40fgh03db811c4',
-            isGroup: true,
-            creator: {
-              _id: '64378abd85008f171cf2990d',
-              firstName: 'Wilt',
-              lastName: 'Shepherd',
-              image: null,
-              email: 'testsuperadmin@example.com',
-              createdAt: '2023-04-13T04:53:17.742Z',
-              __typename: 'User',
-            },
-            organization: {
-              _id: 'pw3ertyuiophgfre45678',
-              name: 'rtyu',
-            },
-            createdAt: '2345678903456',
-            name: 'Test Group Chat',
-            messages: [
-              {
-                _id: '345678',
-                createdAt: '345678908765',
-                messageContent: 'Hello',
-                media: null,
-                replyTo: null,
-                type: 'STRING',
-                sender: {
-                  _id: '2',
-                  firstName: 'Test',
-                  lastName: 'User',
-                  email: 'test@example.com',
-                  image: '',
-                },
-              },
-            ],
-            users: [
-              {
-                _id: '1',
-                firstName: 'Disha',
-                lastName: 'Talreja',
-                email: 'disha@example.com',
-                image: '',
-              },
-              {
-                _id: '2',
-                firstName: 'Test',
-                lastName: 'User',
-                email: 'test@example.com',
-                image: '',
-              },
-              {
-                _id: '3',
-                firstName: 'Test',
-                lastName: 'User1',
-                email: 'test1@example.com',
-                image: '',
-              },
-              {
-                _id: '4',
-                firstName: 'Test',
-                lastName: 'User2',
-                email: 'test2@example.com',
-                image: '',
-              },
-              {
-                _id: '5',
-                firstName: 'Test',
-                lastName: 'User4',
-                email: 'test4@example.com',
-                image: '',
-              },
-            ],
-            admins: [
-              {
-                _id: '1',
-                firstName: 'Disha',
-                lastName: 'Talreja',
-                email: 'disha@example.com',
-                image: '',
-              },
-            ],
-            unseenMessagesByUsers: {
-              '1': 0,
-              '2': 0,
-              '3': 0,
-              '4': 0,
-              '5': 0,
-            },
-          },
-          {
-            _id: '65844efc814ddgh4003db811c4',
-            isGroup: true,
-            creator: {
-              _id: '64378abd85008f171cf2990d',
-              firstName: 'Wilt',
-              lastName: 'Shepherd',
-              image: null,
-              email: 'testsuperadmin@example.com',
-              createdAt: '2023-04-13T04:53:17.742Z',
-              __typename: 'User',
-            },
-            organization: {
-              _id: 'pw3ertyuiophgfre45678',
-              name: 'rtyu',
-            },
-            createdAt: '2345678903456',
-            name: 'Test Group Chat',
-            messages: [
-              {
-                _id: '345678',
-                createdAt: '345678908765',
-                messageContent: 'Hello',
-                media: null,
-                replyTo: null,
-                type: 'STRING',
-                sender: {
-                  _id: '2',
-                  firstName: 'Test',
-                  lastName: 'User',
-                  email: 'test@example.com',
-                  image: '',
-                },
-              },
-            ],
-            users: [
-              {
-                _id: '1',
-                firstName: 'Disha',
-                lastName: 'Talreja',
-                email: 'disha@example.com',
-                image: '',
-              },
-              {
-                _id: '2',
-                firstName: 'Test',
-                lastName: 'User',
-                email: 'test@example.com',
-                image: '',
-              },
-              {
-                _id: '3',
-                firstName: 'Test',
-                lastName: 'User1',
-                email: 'test1@example.com',
-                image: '',
-              },
-              {
-                _id: '4',
-                firstName: 'Test',
-                lastName: 'User2',
-                email: 'test2@example.com',
-                image: '',
-              },
-              {
-                _id: '5',
-                firstName: 'Test',
-                lastName: 'User4',
-                email: 'test4@example.com',
-                image: '',
-              },
-            ],
-            admins: [
-              {
-                _id: '1',
-                firstName: 'Disha',
-                lastName: 'Talreja',
-                email: 'disha@example.com',
-                image: '',
-              },
-            ],
-            unseenMessagesByUsers: {
-              '1': 0,
-              '2': 0,
-              '3': 0,
-              '4': 0,
-              '5': 0,
-            },
-          },
-        ],
-      },
-    },
-  },
-  {
-    request: {
-      query: CHATS_LIST,
-      variables: {
-        id: '',
-      },
-    },
-    result: {
-      data: {
-        chatsByUserId: [
-          {
-            _id: '65844ghjefc814dd4003db811c4',
-            isGroup: true,
-            creator: {
-              _id: '64378abd85008f171cf2990d',
-              firstName: 'Wilt',
-              lastName: 'Shepherd',
-              image: null,
-              email: 'testsuperadmin@example.com',
-              createdAt: '2023-04-13T04:53:17.742Z',
-              __typename: 'User',
-            },
-            organization: {
-              _id: 'pw3ertyuiophgfre45678',
-              name: 'rtyu',
-            },
-            createdAt: '2345678903456',
-            name: 'Test Group Chat',
-            messages: [
-              {
-                _id: '345678',
-                createdAt: '345678908765',
-                messageContent: 'Hello',
-                media: null,
-                replyTo: null,
-                sender: {
-                  _id: '2',
-                  firstName: 'Test',
-                  lastName: 'User',
-                  email: 'test@example.com',
-                  image: '',
-                },
-              },
-            ],
-            users: [
-              {
-                _id: '1',
-                firstName: 'Disha',
-                lastName: 'Talreja',
-                email: 'disha@example.com',
-                image: '',
-              },
-              {
-                _id: '2',
-                firstName: 'Test',
-                lastName: 'User',
-                email: 'test@example.com',
-                image: '',
-              },
-              {
-                _id: '3',
-                firstName: 'Test',
-                lastName: 'User1',
-                email: 'test1@example.com',
-                image: '',
-              },
-              {
-                _id: '4',
-                firstName: 'Test',
-                lastName: 'User2',
-                email: 'test2@example.com',
-                image: '',
-              },
-              {
-                _id: '5',
-                firstName: 'Test',
-                lastName: 'User4',
-                email: 'test4@example.com',
-                image: '',
-              },
-            ],
-            admins: [
-              {
-                _id: '1',
-                firstName: 'Disha',
-                lastName: 'Talreja',
-                email: 'disha@example.com',
-                image: '',
-              },
-            ],
-            unseenMessagesByUsers: {
-              '1': 0,
-              '2': 0,
-              '3': 0,
-              '4': 0,
-              '5': 0,
-            },
-          },
-          {
-            _id: 'ujhgtrdtyuiop',
-            isGroup: true,
-            creator: {
-              _id: '64378abd85008f171cf2990d',
-              firstName: 'Wilt',
-              lastName: 'Shepherd',
-              image: null,
-              email: 'testsuperadmin@example.com',
-              createdAt: '2023-04-13T04:53:17.742Z',
-              __typename: 'User',
-            },
-            organization: {
-              _id: 'pw3ertyuiophgfre45678',
-              name: 'rtyu',
-            },
-            createdAt: '2345678903456',
-            name: 'Test Group Chat',
-            messages: [
-              {
-                _id: '345678',
-                createdAt: '345678908765',
-                messageContent: 'Hello',
-                media: null,
-                replyTo: null,
-                sender: {
-                  _id: '2',
-                  firstName: 'Test',
-                  lastName: 'User',
-                  email: 'test@example.com',
-                  image: '',
-                },
-              },
-            ],
-            users: [
-              {
-                _id: '1',
-                firstName: 'Disha',
-                lastName: 'Talreja',
-                email: 'disha@example.com',
-                image: '',
-              },
-              {
-                _id: '2',
-                firstName: 'Test',
-                lastName: 'User',
-                email: 'test@example.com',
-                image: '',
-              },
-              {
-                _id: '3',
-                firstName: 'Test',
-                lastName: 'User1',
-                email: 'test1@example.com',
-                image: '',
-              },
-              {
-                _id: '4',
-                firstName: 'Test',
-                lastName: 'User2',
-                email: 'test2@example.com',
-                image: '',
-              },
-              {
-                _id: '5',
-                firstName: 'Test',
-                lastName: 'User4',
-                email: 'test4@example.com',
-                image: '',
-              },
-            ],
-            admins: [
-              {
-                _id: '1',
-                firstName: 'Disha',
-                lastName: 'Talreja',
-                email: 'disha@example.com',
-                image: '',
-              },
-            ],
-            unseenMessagesByUsers: {
-              '1': 0,
-              '2': 0,
-              '3': 0,
-              '4': 0,
-              '5': 0,
-            },
-          },
-        ],
-      },
-    },
-  },
-  {
-    request: {
-      query: CHATS_LIST,
-      variables: {
-        id: '1',
-      },
-    },
-    result: {
-      data: {
-        chatsByUserId: [
-          {
-            _id: '65844efc814dhjmkdftyd4003db811c4',
-            isGroup: true,
-            creator: {
-              _id: '64378abd85008f171cf2990d',
-              firstName: 'Wilt',
-              lastName: 'Shepherd',
-              image: null,
-              email: 'testsuperadmin@example.com',
-              createdAt: '2023-04-13T04:53:17.742Z',
-              __typename: 'User',
-            },
-            organization: {
-              _id: 'pw3ertyuiophgfre45678',
-              name: 'rtyu',
-            },
-            createdAt: '2345678903456',
-            name: 'Test Group Chat',
-            messages: [
-              {
-                _id: '345678',
-                createdAt: '345678908765',
-                messageContent: 'Hello',
-                media: null,
-                replyTo: null,
-                sender: {
-                  _id: '2',
-                  firstName: 'Test',
-                  lastName: 'User',
-                  email: 'test@example.com',
-                  image: '',
-                },
-              },
-            ],
-            users: [
-              {
-                _id: '1',
-                firstName: 'Disha',
-                lastName: 'Talreja',
-                email: 'disha@example.com',
-                image: '',
-              },
-              {
-                _id: '2',
-                firstName: 'Test',
-                lastName: 'User',
-                email: 'test@example.com',
-                image: '',
-              },
-              {
-                _id: '3',
-                firstName: 'Test',
-                lastName: 'User1',
-                email: 'test1@example.com',
-                image: '',
-              },
-              {
-                _id: '4',
-                firstName: 'Test',
-                lastName: 'User2',
-                email: 'test2@example.com',
-                image: '',
-              },
-              {
-                _id: '5',
-                firstName: 'Test',
-                lastName: 'User4',
-                email: 'test4@example.com',
-                image: '',
-              },
-            ],
-            admins: [
-              {
-                _id: '1',
-                firstName: 'Disha',
-                lastName: 'Talreja',
-                email: 'disha@example.com',
-                image: '',
-              },
-            ],
-            unseenMessagesByUsers: {
-              '1': 0,
-              '2': 0,
-              '3': 0,
-              '4': 0,
-              '5': 0,
-            },
-          },
-          {
-            _id: '65844ewsedrffc814dd4003db811c4',
-            isGroup: true,
-            creator: {
-              _id: '64378abd85008f171cf2990d',
-              firstName: 'Wilt',
-              lastName: 'Shepherd',
-              image: null,
-              email: 'testsuperadmin@example.com',
-              createdAt: '2023-04-13T04:53:17.742Z',
-              __typename: 'User',
-            },
-            organization: {
-              _id: 'pw3ertyuiophgfre45678',
-              name: 'rtyu',
-            },
-            createdAt: '2345678903456',
-            name: 'Test Group Chat',
-            messages: [
-              {
-                _id: '345678',
-                createdAt: '345678908765',
-                messageContent: 'Hello',
-                media: null,
-                replyTo: null,
-                sender: {
-                  _id: '2',
-                  firstName: 'Test',
-                  lastName: 'User',
-                  email: 'test@example.com',
-                  image: '',
-                },
-              },
-            ],
-            users: [
-              {
-                _id: '1',
-                firstName: 'Disha',
-                lastName: 'Talreja',
-                email: 'disha@example.com',
-                image: '',
-              },
-              {
-                _id: '2',
-                firstName: 'Test',
-                lastName: 'User',
-                email: 'test@example.com',
-                image: '',
-              },
-              {
-                _id: '3',
-                firstName: 'Test',
-                lastName: 'User1',
-                email: 'test1@example.com',
-                image: '',
-              },
-              {
-                _id: '4',
-                firstName: 'Test',
-                lastName: 'User2',
-                email: 'test2@example.com',
-                image: '',
-              },
-              {
-                _id: '5',
-                firstName: 'Test',
-                lastName: 'User4',
-                email: 'test4@example.com',
-                image: '',
-              },
-            ],
-            admins: [
-              {
-                _id: '1',
-                firstName: 'Disha',
-                lastName: 'Talreja',
-                email: 'disha@example.com',
-                image: '',
-              },
-            ],
-            unseenMessagesByUsers: JSON.stringify({
-              '1': 0,
-              '2': 0,
-              '3': 0,
-              '4': 0,
-              '5': 0,
-            }),
-          },
-        ],
-      },
-    },
-  },
-];
-
-export const GROUP_CHAT_BY_ID_QUERY_MOCK = [
-  {
-    request: {
-      query: CHAT_BY_ID,
-      variables: {
-        id: '',
-      },
-    },
-    result: {
-      data: {
-        chatById: {
-          _id: '65844efc814dd4003db811c4',
-          isGroup: true,
-          creator: {
-            _id: '64378abd85008f171cf2990d',
-            firstName: 'Wilt',
-            lastName: 'Shepherd',
-            image: null,
-            email: 'testsuperadmin@example.com',
-            createdAt: '2023-04-13T04:53:17.742Z',
-            __typename: 'User',
-          },
-          organization: {
-            _id: 'pw3ertyuiophgfre45678',
-            name: 'rtyu',
-          },
-          createdAt: '2345678903456',
-          name: 'Test Group Chat',
-          messages: [
-            {
-              _id: '2',
-              createdAt: '345678908765',
-              messageContent: 'Hello',
-              media: null,
-              replyTo: null,
-              sender: {
-                _id: '2',
-                firstName: 'Test',
-                lastName: 'User',
-                email: 'test@example.com',
-                image: '',
-              },
-            },
-          ],
-          users: [
-            {
-              _id: '1',
-              firstName: 'Disha',
-              lastName: 'Talreja',
-              email: 'disha@example.com',
-              image: '',
-            },
-            {
-              _id: '2',
-              firstName: 'Test',
-              lastName: 'User',
-              email: 'test@example.com',
-              image: '',
-            },
-            {
-              _id: '3',
-              firstName: 'Test',
-              lastName: 'User1',
-              email: 'test1@example.com',
-              image: '',
-            },
-            {
-              _id: '4',
-              firstName: 'Test',
-              lastName: 'User2',
-              email: 'test2@example.com',
-              image: '',
-            },
-            {
-              _id: '5',
-              firstName: 'Test',
-              lastName: 'User4',
-              email: 'test4@example.com',
-              image: '',
-            },
-          ],
-          admins: [
-            {
-              _id: '1',
-              firstName: 'Disha',
-              lastName: 'Talreja',
-              email: 'disha@example.com',
-              image: '',
-            },
-          ],
-          unseenMessagesByUsers: JSON.stringify({
-            '1': 0,
-            '2': 0,
-            '3': 0,
-            '4': 0,
-            '5': 0,
-          }),
-        },
-      },
-    },
-  },
-  {
-    request: {
-      query: CHAT_BY_ID,
-      variables: {
-        id: '1',
-      },
-    },
-    result: {
-      data: {
-        chatById: {
-          _id: '65844efc814dd4003db811c4',
-          isGroup: true,
-          creator: {
-            _id: '64378abd85008f171cf2990d',
-            firstName: 'Wilt',
-            lastName: 'Shepherd',
-            image: null,
-            email: 'testsuperadmin@example.com',
-            createdAt: '2023-04-13T04:53:17.742Z',
-            __typename: 'User',
-          },
-          organization: {
-            _id: 'pw3ertyuiophgfre45678',
-            name: 'rtyu',
-          },
-          createdAt: '2345678903456',
-          name: 'Test Group Chat',
-          messages: [
-            {
-              _id: '345678',
-              createdAt: '345678908765',
-              messageContent: 'Hello',
-              media: null,
-              replyTo: null,
-              type: 'STRING',
-              sender: {
-                _id: '2',
-                firstName: 'Test',
-                lastName: 'User',
-                email: 'test@example.com',
-                image: '',
-              },
-            },
-          ],
-          users: [
-            {
-              _id: '1',
-              firstName: 'Disha',
-              lastName: 'Talreja',
-              email: 'disha@example.com',
-              image: '',
-            },
-            {
-              _id: '2',
-              firstName: 'Test',
-              lastName: 'User',
-              email: 'test@example.com',
-              image: '',
-            },
-            {
-              _id: '3',
-              firstName: 'Test',
-              lastName: 'User1',
-              email: 'test1@example.com',
-              image: '',
-            },
-            {
-              _id: '4',
-              firstName: 'Test',
-              lastName: 'User2',
-              email: 'test2@example.com',
-              image: '',
-            },
-            {
-              _id: '5',
-              firstName: 'Test',
-              lastName: 'User4',
-              email: 'test4@example.com',
-              image: '',
-            },
-          ],
-          admins: [
-            {
-              _id: '1',
-              firstName: 'Disha',
-              lastName: 'Talreja',
-              email: 'disha@example.com',
-              image: '',
-            },
-          ],
-          unseenMessagesByUsers: JSON.stringify({
-            '1': 0,
-            '2': 0,
-            '3': 0,
-            '4': 0,
-            '5': 0,
-          }),
-        },
-      },
-    },
-  },
-  {
-    request: {
-      query: CHAT_BY_ID,
-      variables: {
-        id: '1',
-      },
-    },
-    result: {
-      data: {
-        chatById: {
-          _id: '65844efc814dd4003db811c4',
-          isGroup: true,
-          creator: {
-            _id: '64378abd85008f171cf2990d',
-            firstName: 'Wilt',
-            lastName: 'Shepherd',
-            image: null,
-            email: 'testsuperadmin@example.com',
-            createdAt: '2023-04-13T04:53:17.742Z',
-            __typename: 'User',
-          },
-          organization: {
-            _id: 'pw3ertyuiophgfre45678',
-            name: 'rtyu',
-          },
-          createdAt: '2345678903456',
-          name: 'Test Group Chat',
-          messages: [
-            {
-              _id: '345678',
-              createdAt: '345678908765',
-              messageContent: 'Hello',
-              media: null,
-              replyTo: null,
-              type: 'STRING',
-              sender: {
-                _id: '2',
-                firstName: 'Test',
-                lastName: 'User',
-                email: 'test@example.com',
-                image: '',
-              },
-            },
-          ],
-          users: [
-            {
-              _id: '1',
-              firstName: 'Disha',
-              lastName: 'Talreja',
-              email: 'disha@example.com',
-              image: '',
-            },
-            {
-              _id: '2',
-              firstName: 'Test',
-              lastName: 'User',
-              email: 'test@example.com',
-              image: '',
-            },
-            {
-              _id: '3',
-              firstName: 'Test',
-              lastName: 'User1',
-              email: 'test1@example.com',
-              image: '',
-            },
-            {
-              _id: '4',
-              firstName: 'Test',
-              lastName: 'User2',
-              email: 'test2@example.com',
-              image: '',
-            },
-            {
-              _id: '5',
-              firstName: 'Test',
-              lastName: 'User4',
-              email: 'test4@example.com',
-              image: '',
-            },
-          ],
-          admins: [
-            {
-              _id: '1',
-              firstName: 'Disha',
-              lastName: 'Talreja',
-              email: 'disha@example.com',
-              image: '',
-            },
-          ],
-          unseenMessagesByUsers: JSON.stringify({
-            '1': 0,
-            '2': 0,
-            '3': 0,
-            '4': 0,
-            '5': 0,
-          }),
-        },
-      },
-    },
-  },
-  {
-    request: {
-      query: CHAT_BY_ID,
-      variables: {
-        id: '1',
-      },
-    },
-    result: {
-      data: {
-        chatById: {
-          _id: '65844efc814dd4003db811c4',
-          isGroup: true,
-          creator: {
-            _id: '64378abd85008f171cf2990d',
-            firstName: 'Wilt',
-            lastName: 'Shepherd',
-            image: null,
-            email: 'testsuperadmin@example.com',
-            createdAt: '2023-04-13T04:53:17.742Z',
-            __typename: 'User',
-          },
-          organization: {
-            _id: 'pw3ertyuiophgfre45678',
-            name: 'rtyu',
-          },
-          createdAt: '2345678903456',
-          name: 'Test Group Chat',
-          messages: [
-            {
-              _id: '1',
-              createdAt: '345678908765',
-              messageContent: 'Hello',
-              media: null,
-              replyTo: null,
-              sender: {
-                _id: '2',
-                firstName: 'Test',
-                lastName: 'User',
-                email: 'test@example.com',
-                image: '',
-              },
-            },
-          ],
-          users: [
-            {
-              _id: '1',
-              firstName: 'Disha',
-              lastName: 'Talreja',
-              email: 'disha@example.com',
-              image: '',
-            },
-            {
-              _id: '2',
-              firstName: 'Test',
-              lastName: 'User',
-              email: 'test@example.com',
-              image: '',
-            },
-            {
-              _id: '3',
-              firstName: 'Test',
-              lastName: 'User1',
-              email: 'test1@example.com',
-              image: '',
-            },
-            {
-              _id: '4',
-              firstName: 'Test',
-              lastName: 'User2',
-              email: 'test2@example.com',
-              image: '',
-            },
-            {
-              _id: '5',
-              firstName: 'Test',
-              lastName: 'User4',
-              email: 'test4@example.com',
-              image: '',
-            },
-          ],
-          admins: [
-            {
-              _id: '1',
-              firstName: 'Disha',
-              lastName: 'Talreja',
-              email: 'disha@example.com',
-              image: '',
-            },
-          ],
-        },
-      },
-    },
-  },
-  {
-    request: {
-      query: CHAT_BY_ID,
-      variables: {
-        id: '1',
-      },
-    },
-    result: {
-      data: {
-        chatById: {
-          _id: '65844efc814dd4003db811c4',
-          isGroup: true,
-          creator: {
-            _id: '64378abd85008f171cf2990d',
-            firstName: 'Wilt',
-            lastName: 'Shepherd',
-            image: null,
-            email: 'testsuperadmin@example.com',
-            createdAt: '2023-04-13T04:53:17.742Z',
-            __typename: 'User',
-          },
-          organization: {
-            _id: 'pw3ertyuiophgfre45678',
-            name: 'rtyu',
-          },
-          createdAt: '2345678903456',
-          name: 'Test Group Chat',
-          messages: [
-            {
-              _id: '1',
-              createdAt: '345678908765',
-              messageContent: 'Hello',
-              media: '',
-              replyTo: null,
-              sender: {
-                _id: '2',
-                firstName: 'Test',
-                lastName: 'User',
-                email: 'test@example.com',
-                image: '',
-              },
-            },
-          ],
-          users: [
-            {
-              _id: '1',
-              firstName: 'Disha',
-              lastName: 'Talreja',
-              email: 'disha@example.com',
-              image: '',
-            },
-            {
-              _id: '2',
-              firstName: 'Test',
-              lastName: 'User',
-              email: 'test@example.com',
-              image: '',
-            },
-            {
-              _id: '3',
-              firstName: 'Test',
-              lastName: 'User1',
-              email: 'test1@example.com',
-              image: '',
-            },
-            {
-              _id: '4',
-              firstName: 'Test',
-              lastName: 'User2',
-              email: 'test2@example.com',
-              image: '',
-            },
-            {
-              _id: '5',
-              firstName: 'Test',
-              lastName: 'User4',
-              email: 'test4@example.com',
-              image: '',
-            },
-          ],
-          admins: [
-            {
-              _id: '1',
-              firstName: 'Disha',
-              lastName: 'Talreja',
-              email: 'disha@example.com',
-              image: '',
-            },
-          ],
-        },
-      },
-    },
-  },
-];
-
-export const SEND_MESSAGE_TO_CHAT_MOCK = [
-  {
-    request: {
-      query: SEND_MESSAGE_TO_CHAT,
-      variables: {
-        chatId: '1',
-        replyTo: '4',
-        messageContent: 'Test reply message',
-        media: 'https://test.com/image.jpg',
-      },
-    },
-    result: {
-      data: {
-        sendMessageToChat: {
-          _id: '668ec1f1364e03ac47a151',
-          createdAt: '2024-07-10T17:16:33.248Z',
-          messageContent: 'Test ',
-          media: null,
-          replyTo: null,
-          sender: {
-            _id: '64378abd85008f171cf2990d',
-            firstName: 'Wilt',
-            lastName: 'Shepherd',
-            image: '',
-          },
-          updatedAt: '2024-07-10',
-        },
-      },
-    },
-  },
-  {
-    request: {
-      query: SEND_MESSAGE_TO_CHAT,
-      variables: {
-        chatId: '1',
-        replyTo: '4',
-        messageContent: 'Test reply message',
-        media: '',
-      },
-    },
-    result: {
-      data: {
-        sendMessageToChat: {
-          _id: '668ec1f1364e03ac47a151',
-          createdAt: '2024-07-10T17:16:33.248Z',
-          messageContent: 'Test ',
-          media: '',
-          replyTo: null,
-          sender: {
-            _id: '64378abd85008f171cf2990d',
-            firstName: 'Wilt',
-            lastName: 'Shepherd',
-            image: '',
-          },
-          updatedAt: '2024-07-10',
-        },
-      },
-    },
-  },
-  {
-    request: {
-      query: SEND_MESSAGE_TO_CHAT,
-      variables: {
-        chatId: '1',
-        replyTo: '1',
-        messageContent: 'Test reply message',
-      },
-    },
-    result: {
-      data: {
-        sendMessageToChat: {
-          _id: '668ec1f1364e03ac47a151',
-          createdAt: '2024-07-10T17:16:33.248Z',
-          messageContent: 'Test ',
-          media: null,
-          replyTo: null,
-          sender: {
-            _id: '64378abd85008f171cf2990d',
-            firstName: 'Wilt',
-            lastName: 'Shepherd',
-            image: '',
-          },
-          updatedAt: '2024-07-10',
-        },
-      },
-    },
-  },
-  {
-    request: {
-      query: SEND_MESSAGE_TO_CHAT,
-      variables: {
-        chatId: '1',
-        replyTo: undefined,
-        messageContent: 'Hello',
-        media: '',
-      },
-    },
-    result: {
-      data: {
-        sendMessageToChat: {
-          _id: '668ec1f1364e03ac47a151',
-          createdAt: '2024-07-10T17:16:33.248Z',
-          messageContent: 'Test ',
-          media: null,
-          replyTo: null,
-          sender: {
-            _id: '64378abd85008f171cf2990d',
-            firstName: 'Wilt',
-            lastName: 'Shepherd',
-            image: '',
-          },
-          updatedAt: '2024-07-10',
-        },
-      },
-    },
-  },
-  {
-    request: {
-      query: SEND_MESSAGE_TO_CHAT,
-      variables: {
-        chatId: '1',
-        replyTo: '345678',
-        messageContent: 'Test reply message',
-        media: '',
-      },
-    },
-    result: {
-      data: {
-        sendMessageToChat: {
-          _id: '668ec1f1364e03ac47a151',
-          createdAt: '2024-07-10T17:16:33.248Z',
-          messageContent: 'Test ',
-          media: null,
-          replyTo: null,
-          sender: {
-            _id: '64378abd85008f171cf2990d',
-            firstName: 'Wilt',
-            lastName: 'Shepherd',
-            image: '',
-          },
-          updatedAt: '2024-07-10',
-        },
-      },
-    },
-  },
-  {
-    request: {
-      query: SEND_MESSAGE_TO_CHAT,
-      variables: {
-        chatId: '1',
-        replyTo: undefined,
-        messageContent: 'Test message',
-        media: '',
-      },
-    },
-    result: {
-      data: {
-        sendMessageToChat: {
-          _id: '668ec1f1364e03ac47a151',
-          createdAt: '2024-07-10T17:16:33.248Z',
-          messageContent: 'Test ',
-          media: null,
-          replyTo: null,
-          sender: {
-            _id: '64378abd85008f171cf2990d',
-            firstName: 'Wilt',
-            lastName: 'Shepherd',
-            image: '',
-          },
-          updatedAt: '2024-07-10',
-        },
-      },
-    },
-  },
-  {
-    request: {
-      query: SEND_MESSAGE_TO_CHAT,
-      variables: {
-        chatId: '1',
-        replyTo: undefined,
-        messageContent: 'Test reply message',
-        media: '',
-      },
-    },
-    result: {
-      data: {
-        sendMessageToChat: {
-          _id: '668ec1f1364e03ac47a151',
-          createdAt: '2024-07-10T17:16:33.248Z',
-          messageContent: 'Test ',
-          media: null,
-          replyTo: null,
-          sender: {
-            _id: '64378abd85008f171cf2990d',
-            firstName: 'Wilt',
-            lastName: 'Shepherd',
-            image: '',
-          },
-          updatedAt: '2024-07-10',
-        },
-      },
-    },
-  },
-];
-
-export const MARK_CHAT_MESSAGES_AS_READ_MOCK = [
-  {
-    request: {
-      query: MARK_CHAT_MESSAGES_AS_READ,
-      variables: {
-        chatId: '1',
-        userId: '2',
-      },
-    },
-    result: {
-      data: {
-        markChatMessagesAsRead: {
-          _id: 'mocked-id-2',
-          __typename: 'ChatMessage', // Important: __typename needed
-        },
-      },
-    },
-  },
-  {
-    request: {
-      query: MARK_CHAT_MESSAGES_AS_READ,
-      variables: {
-        chatId: '',
-        userId: null,
-      },
-    },
-    result: {
-      data: {
-        markChatMessagesAsRead: {
-          _id: 'mocked-id-null',
-          __typename: 'ChatMessage',
-        },
-      },
-    },
-  },
-  {
-    request: {
-      query: MARK_CHAT_MESSAGES_AS_READ,
-      variables: {
-        chatId: '',
-        userId: 'dummyUser',
-      },
-    },
-    result: {
-      data: {
-        markChatMessagesAsRead: {
-          _id: 'mocked-id-dummy',
-          __typename: 'ChatMessage',
-        },
-      },
-    },
-  },
-  ...Array(5).fill({
-    request: {
-      query: MARK_CHAT_MESSAGES_AS_READ,
-      variables: {
-        chatId: '1',
-        userId: '8',
-      },
-    },
-    result: {
-      data: {
-        markChatMessagesAsRead: {
-          _id: 'mocked-id-8',
-          __typename: 'ChatMessage',
-        },
-      },
-    },
-  }),
-];
